==> old-R/step-00-build-title-categories.R <==
#' ---
#' title: "Title Categories Top 1000"
#' author: "David Selby"
#' date: "6/18/2019"
#' output:
#'  html_document:
#'  toc: true
#'  toc_depth: 1
#' ---
#' 
#' 
#' 
## ----setup, include=FALSE------------------------------------------------
knitr::opts_chunk$set(echo = TRUE)

#' 
#' # Required Packages
#' 
## ------------------------------------------------------------------------
library( tidyverse )
library( reshape2 )
library( dplyr )
library( pander )
library( stringr )
library( hunspell )


#' 
#' # Catagories {.tabset}
#' 
#' Variable Name	        Category Name               
#' --------------       --------------			          
#' CEO	                  Chief Executive		          
#' CFO	                  Financial				            
#' Treasurer	            Treasurer				            
#' DEP.CEO	              Deputy Chief Executive		  
#' SEC	                  Secretary				            
#' COO	                  Operations				          
#' TRUST	                Trustee/Board				        
#' HUM.RES	              Human Resources				      
#' DEP.HEAD	            Department Head				      
#' MAN	Generic           Management				          
#' DEV	                  Development				          
#' TECH	                Technology/IT				        
#' COMM	                Communications				      
#' OTHER	                Other Category				      
#' PROJECT	              Project Manager				      
#' LEGAL	                Legal				           		  
#' FACILITIES            Facilities				          
#' ADMIN.SUP	            Administrative Support		  
#' MED.MAN	              Medical Management				  
#' HEALTH.HUM	          Health and Human Services		
#' TRAIN	                Training/Education		      
#' ACADEMIC.MAN	        Academic Manager				    
#' PROFESIONAL	          Professional				        
#' OTHER.PROF	          Other Professional				  
#' ACADEMIC.PROF	        Academic Professional		    
#' MED.PROF	            Medical Professional			  
#' 
#' 
#' 
#' ## CEO
#' 
## ----CEO-----------------------------------------------------------------
CEO<- c(
" EXECUTIVE DIRECTOR ", 
" PRESIDENT ", 
" PRESIDENT CEO ", 
" CEO ", 
" CHIEF EXECUTIVE OFFICER ", 
" HEAD SCHOOL ", 
" PRESIDENTCEO ", 
" PRINCIPAL ", 
" PRESIDENT DIRECTOR ", 
" FORMER EXECUTIVE DIRECTOR ", 
" HEADMASTER ", 
" CEO PRESIDENT ", 
" PRESIDENT EXECUTIVE DIRECTOR ", 
" EXECUTIVE DIRECTOR CEO ", 
" PRESIDENT CE ", 
" INTERIM EXECUTIVE DIRECTOR ", 
" CO EXECUTIVE DIRECTOR ", 
" EXECUTIVE DIRECTOR ", 
" PRESIDENT ", 
" PRESIDENT CEO ", 
" CEO ", 
" CHIEF EXECUTIVE OFFICER ", 
" HEAD SCHOOL ", 
" PRESIDENTCEO ", 
" PRINCIPAL ", 
" PRESIDENT DIRECTOR ", 
" FORMER EXECUTIVE DIRECTOR ", 
" HEADMASTER ", 
" CEO PRESIDENT ", 
" PRESIDENT EXECUTIVE DIRECTOR ", 
" EXECUTIVE DIRECTOR CEO ", 
" PRESIDENT CE ", 
" INTERIM EXECUTIVE DIRECTOR ", 
" CO EXECUTIVE DIRECTOR ", 
" EXECUTIVE DIRECTOR SECRETARY ", 
" PRESIDENT TREASURER ", 
" SUPERINTENDENT ", 
" SCOUT EXECUTIVE ", 
" SCHOOL DIRECTOR ", 
" FORMER PRESIDENT ", 
" DIRECTOR PRESIDENT ", 
" EXECUTIVE DIRECTOR PRESIDENT ", 
" CEO EXECUTIVE DIRECTOR ", 
" FORMER CEO ", 
" SECRETARY EXECUTIVE DIRECTOR ", 
" FORMER PRESIDENT CEO ", 
" PAST EXECUTIVE DIRECTOR ", 
" PRESIDENT TRUSTEE ", 
" INTERIM CEO ", 
" CHIEF EXECUTIVE ", 
" TRUSTEE PRESIDENT CEO ", 
" CEO DIRECTOR ", 
" DIRECTOR PRESIDENT CEO ", 
" FORMER EXECUTIVE ", 
" PRESIDENT CEO DIRECTOR ", 
" DIRECTOR CEO ", 
" CAMP DIRECTOR ", 
" PRESIDENT SECRETARY ", 
" PRESIDENT COO ", 
" BOARD DIRECTOR ", 
" INTERIM PRESIDENT ", 
" HEAD COACH ", 
" EXECUTIVE DIRECTOR FORMER ", 
" CHIEF ", 
" EXECUTIVE DIRECTOR THRU ", 
" CHAIR CEO ", 
" EXECUTIVE OFFICER ", 
" PRESIDENT CHIEF EXECUTIVE OFFICER ", 
" CEO ADMINISTRATOR ", 
" CEO ADMINISTRATION ", 
" GENERAL DIRECTOR ", 
" VICE PRESIDENT EXECUTIVE DIRECTOR ", 
" CHANCELLOR ", 
" MANAGING ATTORNEY ", 
" TRUSTEE PRESIDENT ", 
" EXECUTIVE DIRECTOR TREASURER ", 
" MUSEUM DIRECTOR ", 
" CEO SECRETARY ", 
" PRESIDENT AN ", 
" EXECUTIVE ", 
" PRESIDENT CHAIR ", 
" EXECUTIVE DIRECTOR UNTIL ", 
" CEOPRESIDENT ", 
" INTERIM EXECUTIVE ", 
" PRESIDENT EX OFFICER ", 
" EXECUTIVE DIRECTOR THROUGH ", 
" ACTING EXECUTIVE DIRECTOR ", 
" PAST PRESIDENT ", 
" CO PRESIDENT ", 
" CHAIR PRESIDENT ", 
" PRESIDENT CHIEF EXECUTIVE OFFICER ", 
" PRESIDENT CEO TRUSTEE ", 
" INTERIM DIRECTOR ", 
" EXECUTIVE DIRECTOR PART YEAR ", 
" GENERAL MGR ", 
" PRESIDENT CEO EX OFFICER ", 
" PRESIDENT CHIEF EXECUTIVE ", 
" EXECUTIVE DIRECTOR EX OFFICER ", 
" PRESIDENTDIRECTOR ", 
" PRESIDENTEXECUTIVE DIRECTOR ", 
" PRIOR EXECUTIVE DIRECTOR ", 
" PRESIDENT EMERITUS ", 
" PRESIDENT EX ", 
" PRESIDENT FOUNDER ", 
" DIRECTOR EXECUTIVE DIRECTOR ", 
" CHIEF EXECUTIVE O ", 
" EXECUTIVE PRODUCER ", 
" FOUNDER CEO ", 
" MIDDLE SCHOOL PRINCIPAL ", 
" SCHOOL PRINCIPAL ", 
" UPPER SCHOOL HEAD ", 
" EXECUTIVE DIRECTOR PHYSICIAN ", 
" BOARD MEMBER EXECUTIVE DIRECTOR ", 
" EDITOR IN CHIEF ", 
" BOARD MEMBER CEO ", 
" INTERIM PRESIDENT CEO ", 
" FORMER HEAD SCHOOL ", 
" EXECUTIVE DIRECTOR VICE PRESIDENT ", 
" PRESIDENT CEO SECRETARY ", 
" FOUNDATION PRESIDENT ", 
" SECRETARY CEO ", 
" ADMINISTRATOR CEO ", 
" EXECUTIVE VICE PRESIDENT CEO ", 
" SCOUT EXECUTIVE CEO ", 
" CEO BOARD MEMBER ", 
" CHIEF EXECUTIVE DIRECTOR ", 
" EXECUTIVE DIRECTOR", 
" FOUNDER PRESIDENT ", 
" CHAPTER PRESIDENT ", 
" CEO THROUGH ", 
" CEO THRU ", 
" MIDDLE SCHOOL HEAD ", 
" EX OFFICER DIRECTOR ", 
" EXECUTIVE DIRECTOR DIRECTOR ", 
" REGIONAL EXECUTIVE DIRECTOR ", 
" FORMER CHIEF EXECUTIVE OFFICER ", 
" HIGH SCHOOL PRINCIPAL ", 
" LOWER SCHOOL HEAD ", 
" CEO FOUNDER ", 
" CHAIR DIRECTOR ", 
" EXECUTIVE DIRECTOR BOARD MEMBER ", 
" EXECUTIVE EDITOR ", 
" HEAD MASTER ", 
" FOUNDER EXECUTIVE DIRECTOR ", 
" MIDDLE SCHOOL DIRECTOR ", 
" PRESIDENT CEO THRU ", 
" TREASURER EXECUTIVE DIRECTOR ", 
" PRINCIPAL DIRECTOR ", 
" UPPER SCHOOL DIRECTOR ", 
" CEO CHAIR ", 
" CEO PRINCIPAL ", 
" HOSPITAL PRESIDENT ", 
" PRESIDENTTREASURER ", 
" PRESIDENT TREASURER DIRECTOR ", 
" DIRECTOR UPPER SCHOOL ", 
" EXECUTIVE DIRECTOR TRUSTEE ", 
" PRESIDENTCHIEF EXECUTIVE OFFICER ", 
" TRUSTEE EXECUTIVE DIRECTOR ", 
" UPPER SCHOOL PRINCIPAL ", 
" CO CEO ", 
" DIRECTOR MIDDLE SCHOOL ", 
" INTERIM HEAD SCHOOL ", 
" LOWER SCHOOL PRINCIPAL ", 
" PRESIDENT CPO ", 
" CURRENT EXECUTIVE DIRECTOR ", 
" VICE CHAIR CEO ", 
" EXECUTIVE DIRECTOR FROM ", 
" UNIVERSITY PRESIDENT ", 
" CEO TRUSTEE ", 
" DIRECTOR LOWER SCHOOL ", 
" EXECUTIVE DIRECTOR CFO ", 
" EXECUTIVE DIRECTOR CURRENT ", 
" EXECUTIVE DIRECTOR FOUNDATION ", 
" LOWER SCHOOL DIRECTOR ", 
" PAST CEO ", 
" PRESIDENTTRUSTEE ", 
" PRESIDENT C ", 
" CEO FORMER ", 
" FORMER PRESIDENTCEO ", 
" PRESIDENT THRU ", 
" BOARD MEMBER PRESIDENT CEO ", 
" EXECUTIVE DIRECTOR AS ", 
" EXECUTIVE DIRECTOR RETIRED ", 
" PRESIDENT THROUGH ", 
" BRANCH EXECUTIVE DIRECTOR ", 
" DIRECTOR CHAIR ", 
" PRESIDENT ADMINISTRATOR ", 
" PRESIDENT CEO UNTIL ", 
" BOARD MEMBER PRESIDENT ", 
" EXECUTIVE DIRECTOR PARTIAL YEAR ", 
" BOARD PRESIDENT ", 
" PRESIDENT FOUNDATION ", 
" HEAD SCHOOLS ", 
" PRESIDENT CEO BOARD MEMBER ", 
" COLLEGE PRESIDENT ", 
" DIRECTOR PRINCIPAL ", 
" EXECDIRECTOR ", 
" PRESIDENT CFO ", 
" PRESIDENT FORMER ", 
" PRESIDENT UNTIL ", 
" SCHOOL LEADER ", 
" SENIOR VICE PRESIDENT SECRETARY ",
" CEO",
" PRESIDENT CEO") 

CEO <- unique(CEO)


#' 
#' 
#' ## CEO Problems
#' 
#' HEAD SCHOOL
#' 
#' CAMP DIRECTOR
#' GENERAL MGR
#' 
#' 
## ----CEO Problems--------------------------------------------------------

CEO.Prob <- c(
" SCOUT EXECUTIVE ", 
" SCHOOL DIRECTOR ", 
" BOARD DIRECTOR ", 
" HEAD COACH ", 
" CHIEF ", 
" GENERAL DIRECTOR ", 
" VICE PRESIDENT EXECUTIVE DIRECTOR ", 
" MANAGING ATTORNEY ", 

" MUSEUM DIRECTOR ", 

" INTERIM DIRECTOR ", 
" PRESIDENT CEO EX OFFICER ", 
" PRESIDENT FOUNDER ", 
" MIDDLE SCHOOL PRINCIPAL ", 
" SCHOOL PRINCIPAL ", 
" UPPER SCHOOL HEAD ", 
" BOARD MEMBER EXECUTIVE DIRECTOR ", 
" EDITOR IN CHIEF ", 
" FORMER HEAD SCHOOL ", 
" CHAPTER PRESIDENT ", 
" MIDDLE SCHOOL HEAD ", 
" EX OFFICER DIRECTOR ", 
" REGIONAL EXECUTIVE DIRECTOR ", 
" HIGH SCHOOL PRINCIPAL ", 
" LOWER SCHOOL HEAD ", 
" CHAIR DIRECTOR ", 
" EXECUTIVE DIRECTOR BOARD MEMBER ", 
" EXECUTIVE EDITOR ", 
" HEAD MASTER ", 
" MIDDLE SCHOOL DIRECTOR ", 
" PRINCIPAL DIRECTOR ", 
" UPPER SCHOOL DIRECTOR ", 
" CEO CHAIR ", 
" DIRECTOR UPPER SCHOOL ", 

" UPPER SCHOOL PRINCIPAL ", 
" DIRECTOR MIDDLE SCHOOL ", 
" INTERIM HEAD SCHOOL ", 
" LOWER SCHOOL PRINCIPAL ", 

" DIRECTOR LOWER SCHOOL ", 
" LOWER SCHOOL DIRECTOR ", 
" BOARD MEMBER PRESIDENT CEO ", 
" EXECUTIVE DIRECTOR RETIRED ", 
" BRANCH EXECUTIVE DIRECTOR ", 
" DIRECTOR CHAIR ", 
" PRESIDENT ADMINISTRATOR ", 
" BOARD MEMBER PRESIDENT ", 
" BOARD PRESIDENT ", 
" HEAD SCHOOLS ", 
" COLLEGE PRESIDENT ", 
" DIRECTOR PRINCIPAL ", 
" SCHOOL LEADER ", 
" SENIOR VICE PRESIDENT SECRETARY ",
" EXECUTIVE DIRECTOR BOARD MEMBER ", 

" PRESIDENT CEO BOARD MEMBER ",
" HEAD SCHOOL ",
" HEADMASTER ",
" CAMP DIRECTOR ",
" GENERAL MGR ",
" MS PRINCIPAL "
)
CEO.Prob <- unique(CEO.Prob)

CEO.Board <-c(
" PRESIDENTTRUSTEE ", 
" EXECUTIVE DIRECTOR TRUSTEE ", 
" TRUSTEE EXECUTIVE DIRECTOR ", 
" VICE CHAIR CEO ", 
" CEO TRUSTEE ", 
" TRUSTEE PRESIDENT ", 
" PRESIDENT CEO TRUSTEE ") %>% unique



CEO.SCHOOL <- c(
" HEAD COACH ",
" MIDDLE SCHOOL PRINCIPAL ",
" SCHOOL PRINCIPAL ",
" UPPER SCHOOL HEAD ",
" MIDDLE SCHOOL HEAD ",
" REGIONAL EXECUTIVE DIRECTOR ",
" HIGH SCHOOL PRINCIPAL ",
" LOWER SCHOOL HEAD ",
" HEAD MASTER ",
" UPPER SCHOOL DIRECTOR ",
" DIRECTOR UPPER SCHOOL ",
" UPPER SCHOOL PRINCIPAL ",
" DIRECTOR MIDDLE SCHOOL ",
" INTERIM HEAD SCHOOL ",
" LOWER SCHOOL PRINCIPAL ",
" DIRECTOR LOWER SCHOOL ",
" LOWER SCHOOL DIRECTOR ",
" HEAD SCHOOLS ",
" SCHOOL LEADER ",
" SCHOOL DIRECTOR ",
" FORMER HEAD SCHOOL ",
" MIDDLE SCHOOL DIRECTOR ")

SCHOOL.HEAD <- c(
" ASSISTANT HEAD SCHOOL ",
" HEAD SCHOOL ",
" HEAD UPPER SCHOOL ",
" HEAD LOWER SCHOOL ",
" HEAD MIDDLE SCHOOL ",
" ASSOCIATE HEAD SCHOOL ",
" FORFORMER HEAD SCHOOL ",
" DIRECTOR COACHING ",
" DIRECTOR UPPER SCHOOL ",
" DIRECTOR LOWER SCHOOL ",
" DIRECTOR MIDDLE SCHOOL ",
" INTERIM HEAD SCHOOL "
)
CEO.Prob.Other <- " MANAGING ATTORNEY "



#' 
#' ## CFO
#' 
## ----CFO-----------------------------------------------------------------
CFO <- c( " CFO ", 
" CHIEF FINANCIAL OFFICER ", 
" DIRECTOR FINANCE ", 
" FINANCE DIRECTOR ", 
" VICE PRESIDENT FINANCE ", 
" VICE PRESIDENT CFO ", 
" SENIOR VICE PRESIDENT CFO ", 
" CFO TREASURER ", 
" CFO ", 
" CHIEF FINANCIAL OFFICER ", 
" DIRECTOR FINANCE ", 
" FINANCE DIRECTOR ", 
" VICE PRESIDENT FINANCE ", 
" VICE PRESIDENT CFO ", 
" SENIOR VICE PRESIDENT CFO ", 
" CFO TREASURER ", 
" EXECUTIVE VICE PRESIDENT CFO ", 
" TREASURER CFO ", 
" FINANCE MANAGER ", 
" VICE PRESIDENT FINANCE CFO ", 
" FISCAL DIRECTOR ", 
" VICE PRESIDENT FINANCE ADMINISTRATION ", 
" DIRECTOR FINANCE ADMINISTRATION ", 
" FORMER CFO ", 
" FISCAL OFFICER ", 
" FINANCIAL DIRECTOR ", 
" COMPTROLLER ", 
" CFO COO ", 
" DIRECTOR FINANCE OPERATIONS ", 
" ACCOUNTANT ", 
" VICE PRESIDENT FINANCE CFO ", 
" COO CFO ", 
" DIRECTOR FINANCE ", 
" BOOKKEEPER ", 
" DIRECTOR FINANCEOPERATIONS ", 
" VICE PRESIDENT FINANCE ADMINISTRATION ", 
" FINANCE OFFICER ", 
" SENIOR VICE PRESIDENT FINANCE CFO ", 
" SENIOR VICE PRESIDENT FINANCE ", 
" CHIEF FINANCE ", 
" VICE PRESIDENT CHIEF FINANCIAL OFFICER ", 
" FINANCIAL MANAGER ", 
" INTERIM CFO ", 
" VICE PRESIDENT FINANCE ", 
" VICE PRESIDENTCFO ", 
" FISCAL MANAGER ", 
" DIRECTOR FINANCEADMINISTRATION ", 
" DIRECTOR ACCOUNTING ", 
" FINANCIAL OFFICER ", 
" ACCOUNTING MANAGER ", 
" ASSISTANT TREASURER CFO ", 
" FINANCE OFFICER ", 
" CFO ASSISTANT TREASURER ", 
" VICE PRESIDENT FINANCE OPERATIONS ", 
" VICE PRESIDENT FINANCE TREASURER ", 
" CORPORATE CONTROLLER ", 
" DIRECTOR FINANCE ADMINISTRATION ", 
" SENIOR VICE PRESIDENT CHIEF FINANCIAL OFFICER ", 
" VICE PRESIDENT FINANCEADMINISTRATION ", 
" VICE PRESIDENT BUSINESS FINANCE ", 
" CFO VICE PRESIDENT FINANCE ", 
" VICE PRESIDENT FINANCIAL SERVICES ", 
" CHIEF FINANCE OFFICER ", 
" CHIEF FISCAL OFFICER ", 
" SENIOR VICE PRESIDENT CFO TREASURER ", 
" SECRETARY TREASURER CFO ", 
" VICE PRESIDENT REVENUE CYCLE ", 
" VICE PRESIDENT FINANCE ADMINISTRATION ", 
" CFO THRU ", 
" CFO VICE PRESIDENT FINANCE ", 
" SENIOR VICE PRESIDENTCFO ", 
" FORMER CHIEF FINANCIAL OFFICER ", 
" CFO THROUGH ", 
" SENIOR DIRECTOR FINANCE ", 
" TREASURERCFO ", 
" CFO VICE PRESIDENT ", 
" VICE PRESIDENT FINANCEADMINISTRATION ", 
" CEO CFO ", 
" SENIOR VICE PRESIDENT FINANCE ADMINISTRATION ", 
" DIRECTOR FINANCE OPERATIONS ", 
" VICE PRESIDENT CFO TREASURER ", 
" FINANCE MANAGEMENT ", 
" EXECUTIVE VICE PRESIDENTCFO ", 
" TREASURER CF ", 
" CFO UNTIL ", 
" CFO SENIOR VICE PRESIDENT ", 
" DIRECTOR FINANCE HUMAN RESOURCES ", 
" DIRECTOR ADMINISTRATION FINANCE ", 
" EXECUTIVE VICE PRESIDENT FINANCE CFO ", 
" SENIOR VICE PRESIDENT FINANCE CFO ", 
" ACCOUNTING DIRECTOR ", 
" SECRETARY CFO ", 
" TREASURER VICE PRESIDENT FINANCE ", 
" VICE PRESIDENT FINANCE OPERATIONS ", 
" CFO FORMER ", 
" EXECUTIVE VICE PRESIDENT CFO TREASURER ", 
" EXECUTIVE DIRECTOR CFO ", 
" FINANCIAL ", 
" DIRECTOR BUSINESS FINANCE ", 
" FISCAL MANAGEMENT ", 
" TREASURER SENIOR VICE PRESIDENT CFO ", 
" CFO SECRETARY ", 
" FORMER DIRECTOR FINANCE ", 
" SENIOR ACCOUNTANT ", 
" EXECUTIVE VICE PRESIDENT CHIEF FINANCIAL OFFICER ", 
" ASSISTANT VICE PRESIDENT FINANCE ", 
" CONTROLLER CFO ", 
" CFO CONTROLLER ", 
" CFOTREASURER ", 
" CFO PART YEAR ", 
" CFO SECRETARY TREASURER ", 
" CHIEF FINANCIAL OFFICER TREASURER ", 
" VICE PRESIDENT BUSINESS FINANCE ", 
" PRESIDENT CFO ", 
" SENIOR VICE PRESIDENT FINANCE ",
" DIRECTOR ACCOUNTING "
 )

CFO <- unique(CFO)


#' 
#' ## CFO Problems
#' 
## ----CFO Problems--------------------------------------------------------
CFO.Prob <- c(" SENIOR VICE PRESIDENT FINANCE ADMINISTRATION ", " TREASURER CF ")

CFO.Prob <- unique(CFO.Prob)

#' 
#' ## Treasurer
#' 
## ----Treasurer-----------------------------------------------------------

Treasurer <- c( 
" TREASURER ", 
" SECRETARY TREASURER ", 
" ASSISTANT SECRETARY TREASURER ", 
" ASSISTANT TREASURER ", 
" CFO TREASURER ", 
" TREASURER ", 
" PRESIDENT TREASURER ", 
" SECRETARY TREASURER ", 
" ASSISTANT SECRETARY TREASURER ", 
" ASSISTANT TREASURER ", 
" CFO TREASURER ", 
" TREASURER CFO ", 
" VICE PRESIDENT TREASURER ", 
" TREASURER SECRETARY ", 
" EXECUTIVE DIRECTOR TREASURER ", 
" ASSISTANT TREASURER CFO ", 
" CFO ASSISTANT TREASURER ", 
" TREASURER DIRECTOR ", 
" VICE PRESIDENT FINANCE TREASURER ", 
" ASSISTANT SECRETARY ASSISTANT TREASURER ", 
" FORMER ASSISTANT SECRETARY TREASURER ", 
" SENIOR VICE PRESIDENT CFO TREASURER ", 
" SECRETARY TREASURER CFO ", 
" SENIOR VICE PRESIDENT TREASURER ", 
" DIRECTOR TREASURER ", 
" SECRETARY TREASURER DIRECTOR ", 
" TREASURERCFO ", 
" VICE PRESIDENT CFO TREASURER ", 
" EXECUTIVE DIRECTOR SECRETARY TREASURER ", 
" TREASURER CF ", 
" EXECUTIVE VICE PRESIDENT TREASURER ", 
" TREASURER EXECUTIVE DIRECTOR ", 
" PRESIDENTTREASURER ", 
" PRESIDENT TREASURER DIRECTOR ", 
" TREASURER VICE PRESIDENT FINANCE ", 
" EXECUTIVE VICE PRESIDENT CFO TREASURER ", 
" TREASURER SENIOR VICE PRESIDENT CFO ", 
" TREASURER TEACHER ", 
" CONTROLLER ASSISTANT TREASURER ", 
" CFOTREASURER ", 
" CFO SECRETARY TREASURER ", 
" CHIEF FINANCIAL OFFICER TREASURER ", 
" SECY TREASURER " )

Treasurer <- Treasurer %>% unique


#' 
#' ## Treasurer Problems
#' 
## ----Treasurer Problems--------------------------------------------------
Treasurer.Prob <- c(" TREASURER TEACHER " )

Treasurer.Prob <- Treasurer.Prob %>% unique

#' 
#' ## Deputy CEO
#' 
## ----Deputry CEO---------------------------------------------------------

DEP.CEO <- c(
" DEPUTY DIRECTOR ", 
" ASSOCIATE EXECUTIVE DIRECTOR ", 
" ASSISTANT EXECUTIVE DIRECTOR ", 
" ASSISTANT EXECUTIVE DIRECTOR ", 
" DEPUTY DIRECTOR ", 
" ASSOCIATE EXECUTIVE DIRECTOR ", 
" ASSISTANT EXECUTIVE DIRECTOR ", 
" DEPUTY EXECUTIVE DIRECTOR ", 
" ASSISTANT HEAD SCHOOL ", 
" ASSOCIATE HEAD SCHOOL ", 
" ASSISTANT HEADMASTER ", 
" DEPUTY CEO ", 
" ASSISTANT CHIEF ", 
" DEPUTY EXECUTIVE ", 
" CFO SECRETARY TREASURER ", 
" EXECUTIVE VICE PRESIDENT PROVOST ") 

DEP.CEO <- DEP.CEO %>% unique



#' 
#' 
#' ## Deputy CEO Problems
#' 
## ----Deputy CEO Problems-------------------------------------------------
DEP.CEO.Prob <- c( " ASSISTANT CHIEF ", " DEPUTY EXECUTIVE " )

DEP.CEO.Prob <- DEP.CEO.Prob %>% unique


#' 
#' 
#' ## Secretary 
#' 
## ----Secretary-----------------------------------------------------------
SEC <- c( 
" SECRETARY ", 
" ASSISTANT SECRETARY ", 
" SECRETARY TREASURER ", 
" ASSISTANT SECRETARY TREASURER ", 
" SECRETARY ", 
" ASSISTANT SECRETARY ", 
" SECRETARY TREASURER ", 
" ASSISTANT SECRETARY TREASURER ", 
" EXECUTIVE DIRECTOR SECRETARY ", 
" SECRETARY EXECUTIVE DIRECTOR ", 
" CORPORATE SECRETARY ", 
" VICE PRESIDENT SECRETARY ", 
" SECRETARY DIRECTOR ", 
" PRESIDENT SECRETARY ", 
" EXECUTIVE SECRETARY ", 
" TREASURER SECRETARY ", 
" BOARD SECRETARY ", 
" CEO SECRETARY ", 
" DIRECTOR SECRETARY ", 
" ASSISTANT SECRETARY ASSISTANT TREASURER ", 
" GENERAL COUNSEL SECRETARY ", 
" FORMER ASSISTANT SECRETARY TREASURER ", 
" SECRETARY TREASURER CFO ", 
" SECRETARY GENERAL COUNSEL ", 
" SECRETARY TEACHER ", 
" SCOUT EXECUTIVE SECRETARY ", 
" SECRETARY TREASURER DIRECTOR ", 
" PRESIDENT CEO SECRETARY ", 
" SECRETARY CEO ", 
" VICE PRESIDENT ASSISTANT SECRETARY ", 
" EXECUTIVE DIRECTOR SECRETARY TREASURER ", 
" SECRETARY COO ", 
" COO SECRETARY ", 
" SECRETARY CFO ", 
" CORP SECRETARY ", 
" SECRETARY VICE PRESIDENT ", 
" CFO SECRETARY ", 
" SECY TREASURER ", 
" RECORDING SECRETARY ", 
" SENIOR VICE PRESIDENT SECRETARY ")

SEC <- SEC %>% unique

#' 
#' ## Secretary Problems
#' 
## ----Secretary Problems--------------------------------------------------
SEC.Prob <- c( " SENIOR VICE PRESIDENT SECRETARY ")



#' 
#' ## COO
#' 
## ----COO-----------------------------------------------------------------
COO <- c( " COO ", 
" CHIEF OPERATING OFFICER ", 
" MANAGING DIRECTOR ", 
" DIRECTOR OPERATIONS ", 
" VICE PRESIDENT OPERATIONS ", 
" PROVOST ", 
" EXECUTIVE VICE PRESIDENT COO ", 
" VICE PRESIDENT OPERATIONS ", 
" CMO ", 
" VICE PRESIDENT COO ", 
" CHIEF OPERATIONS OFFICER ", 
" INTERIM PROVOST ", 
" COO ", 
" CHIEF OPERATING OFFICER ", 
" MANAGING DIRECTOR ", 
" DIRECTOR OPERATIONS ", 
" VICE PRESIDENT OPERATIONS ", 
" PROVOST ", 
" EXECUTIVE VICE PRESIDENT COO ", 
" VICE PRESIDENT OPERATIONS ", 
" CMO ", 
" VICE PRESIDENT COO ", 
" CHIEF OPERATIONS OFFICER ", 
" CHIEF STAFF ", 
" OPERATIONS DIRECTOR ", 
" VICE PRESIDENT FINANCE ADMINISTRATION ", 
" DIRECTOR FINANCE ADMINISTRATION ", 
" SENIOR VICE PRESIDENT COO ", 
" CHIEF OPERATIONS ", 
" CFO COO ", 
" ADMINISTRATIVE DIRECTOR ", 
" SENIOR VICE PRESIDENT OPERATIONS ", 
" ADMINISTRATION DIRECTOR ", 
" DIRECTOR ADMINISTRATION ", 
" OPERATIONS MANAGER ", 
" ADMINISTRATION ", 
" COO CFO ", 
" DIRECTOR OPERATIONS ", 
" DIRECTOR FINANCEOPERATIONS ", 
" CHIEF ADMINISTRATION OFFICER ", 
" PRESIDENT COO ", 
" VICE PRESIDENT ADMINISTRATION ", 
" VICE PRESIDENT FINANCE ADMINISTRATION ", 
" CHIEF OPERATIONS ", 
" SENIOR VICE PRESIDENT OPERATIONS ", 
" SENIOR VICE PRESIDENT CMO ", 
" DIRECTOR FINANCEADMINISTRATION ", 
" VICE PRESIDENT CMO ", 
" VICE PRESIDENT CLINICAL OPERATIONS ", 
" VICE PRESIDENT FINANCEADMINISTRATION ", 
" VICE PRESIDENT ADMINISTRATION FINANCE ", 
" VICE PRESIDENT FINANCEADMINISTRATION ", 
" DIRECTOR FINANCE ADMINISTRATION ", 
" VICE PRESIDENT CONTROLLER ", 
" EXECUTIVE VICE PRESIDENT CHIEF OPERATING OFFICER ", 
" EXECUTIVE VICE PRESIDENTCOO ", 
" DIRECTOR BUSINESS OPERATIONS ", 
" GENERAL MANAGEMENT ", 
" SENIOR VICE PRESIDENT ADMINISTRATION ", 
" OPERATIONS D ", 
" VICE PRESIDENT ADMINISTRATOR ", 
" VICE PRESIDENT FINANCEADMINISTRATION ", 
" SENIOR VICE PRESIDENT FINANCE ADMINISTRATION ", 
" CHIEF ADMINISTRATION ", 
" FORMER COO ", 
" OPERATIONS ", 
" OPERATIONS MGR ", 
" VICE PRESIDENT CHIEF OPERATING OFFICER ", 
" SENIOR VICE PRESIDENTCOO ", 
" EXECUTIVE DIRECTOR COO ", 
" EXECUTIVE DIRECTOR OPERATIONS ", 
" SENIOR VICE PRESIDENT CHIEF OPERATING OFFICER ", 
" EXECUTIVE VICE PRESIDENT OPERATIONS ", 
" VICE PRESIDENT ADMINISTRATION ", 
" ASSISTANT ADMINISTRATION ", 
" MANAGING EDITOR ", 
" VICE PRESIDENTCOO ", 
" DIRECTOR ADMINISTRATION FINANCE ", 
" BUSINESS ADMINISTRATION ", 
" SECRETARY COO ", 
" COO SECRETARY ", 
" EXECUTIVE ADMINISTRATOR ", 
" OPERATIONS M ", 
" DIRECTOR FINANCEAD ", 
" COO VICE PRESIDENT ", 
" EXECUTIVE VICE PRESIDENT CMO ", 
" PRESIDENT ADMINISTRATOR ", 
" EXECUTIVE VICE PRESIDENT OPERATIONS ", 
" SENIOR DIRECTOR OPERATIONS ", 
" CHIEF OPERATIONAL OFFICER ", 
" EXECUTIVE DIRECTOR OPERATIONS ", 
" VICE PRESIDENT HOSPITAL OPERATIONS ", 
" VICE PRESIDENT OPERATIONS ", 
" INTERIM PROVOST ",
" VICE PRESIDENT ADMINISTRATION ",
" CHIEF ADMINISTRATIVE OFFICER ",
" DIRECTOR ADMINISTRATION "


 )

COO <- COO %>% unique


#' 
#' ## COO Problems
#' 
## ----COO Problems--------------------------------------------------------



COO.Prob <- c( 
" PROVOST ", 
" INTERIM PROVOST ", 
" SENIOR VICE PRESIDENT CMO ", 
" SENIOR VICE PRESIDENT FINANCE ADMINISTRATION ", 
" CHIEF ADMINISTRATION ", 
" MANAGING EDITOR ", 
" EXECUTIVE ADMINISTRATOR ", 
" PRESIDENT ADMINISTRATOR ", 
" INTERIM PROVOST ",
" NURSING HOME ADMINISTRATOR "
 )

COO.Prob <- COO.Prob %>% unique


#' 
#' ## Director/Trustee 
#' 
## ----Director/Trustee----------------------------------------------------

TRUST <- c(
" BOARD MEMBER ", 
" TRUSTEE ", 
" CHAIR ", 
" MEMBER ", 
" BOARD MEMBER ", 
" TRUSTEE ", 
" CHAIR ", 
" MEMBER ", 
" VICE CHAIR ", 
" FOUNDATION DIRECTOR ", 
" PRESIDENT TRUSTEE ", 
" BOARD MEMBER PHYSICIAN ", 
" BOARD MEMBER TEACHER ", 
" COMMISSIONER ", 
" MEMBER TEACHER ", 
" TRUSTEE PRESIDENT CEO ", 
" SENIOR FELLOW ", 
" PHYSICIAN BOARD MEMBER ", 
" TRUSTEE PHYSICIAN ", 
" FOUNDER ", 
" TRUSTEE PRESIDENT ", 
" PRESIDENT CHAIR ", 
" CHIEF PARTY ", 
" CHAIR PRESIDENT ", 
" PRESIDENT CEO TRUSTEE ", 
" PARTNER ", 
" TRUSTEE TEACHER ", 
" BOARD CHAIR ", 
" FACULTY TRUSTEE ", 
" PRESIDENT BOARD MEMBER ", 
" BOARD MEMBER EXECUTIVE DIRECTOR ", 
" FOUNDING DIRECTOR ", 
" BOARD MEMBER CEO ", 
" BOARD MEMBERDEPT CHAIR ", 
" TEACHER BOARD MEMBER ", 
" TRUSTEE FACULTY ", 
" PHYSICIAN TRUSTEE ", 
" BOARD DIRECTOR ", 
" FACULTY REPRESENTATIVE ", 
" CHAIR DIRECTOR ", 
" EXECUTIVE DIRECTOR BOARD MEMBER ", 
" CO FOUNDER ", 
" CHAIR THE BOARD ", 
" CEO CHAIR ", 
" FOUNDER DIRECTOR ", 
" EXECUTIVE DIRECTOR TRUSTEE ", 
" TRUSTEE EXECUTIVE DIRECTOR ", 
" VICE CHAIR CEO ", 
" CEO TRUSTEE ", 
" PRESIDENTTRUSTEE ", 
" BOARD MEMBER PRESIDENT CEO ", 
" ASSOCIATE COMMISSIONER ", 
" DIRECTOR CHAIR ", 
" MEDICAL STAFF PRESIDENT ", 
" BOARD MEMBER PRESIDENT ", 
" BOARD PRESIDENT ", 
" EX OFFICER MEMBER ", 
" PRESIDENT CEO BOARD MEMBER ", 
" STAFF REPRESENTATIVE ",
" CHIEF STAFF "
)

TRUST <- TRUST %>% unique


#' 
#' 
#' ## Director/Trustee Problems
#' 
## ----Director/Trustee Problems-------------------------------------------
TRUST.Prob <- c( 
" CHAIR ", 
" MEMBER ", 
" CHAIR ", 
" MEMBER ", 
" VICE CHAIR ", 
" FOUNDATION DIRECTOR ", 
" PRESIDENT TRUSTEE ", 
" COMMISSIONER ", 
" MEMBER TEACHER ", 
" TRUSTEE PRESIDENT CEO ", 
" SENIOR FELLOW ", 
" FOUNDER ", 
" TRUSTEE PRESIDENT ", 
" PRESIDENT CHAIR ", 
" CHIEF PARTY ", 
" CHAIR PRESIDENT ", 
" PRESIDENT CEO TRUSTEE ", 
" PARTNER ", 
" BOARD CHAIR ", 
" PRESIDENT BOARD MEMBER ", 
" BOARD MEMBER EXECUTIVE DIRECTOR ", 
" FOUNDING DIRECTOR ", 
" BOARD MEMBER CEO ", 
" BOARD MEMBERDEPT CHAIR ", 
" BOARD DIRECTOR ", 
" FACULTY REPRESENTATIVE ", 
" CHAIR DIRECTOR ", 
" EXECUTIVE DIRECTOR BOARD MEMBER ", 
" CO FOUNDER ", 
" CHAIR THE BOARD ", 
" CEO CHAIR ", 
" FOUNDER DIRECTOR ", 
" EXECUTIVE DIRECTOR TRUSTEE ", 
" TRUSTEE EXECUTIVE DIRECTOR ", 
" VICE CHAIR CEO ", 
" CEO TRUSTEE ", 
" PRESIDENTTRUSTEE ", 
" BOARD MEMBER PRESIDENT CEO ", 
" ASSOCIATE COMMISSIONER ", 
" DIRECTOR CHAIR ", 
" MEDICAL STAFF PRESIDENT ", 
" BOARD MEMBER PRESIDENT ", 
" BOARD PRESIDENT ", 
" EX OFFICER MEMBER ", 
" PRESIDENT CEO BOARD MEMBER ", 
" STAFF REPRESENTATIVE " )

TRUST.Prob <- TRUST.Prob %>% unique


TRUST.Prob.CEO <- c(
" CHAIR ",
" FOUNDATION DIRECTOR ",
" PRESIDENT TRUSTEE ",
" COMMISSIONER ",
" TRUSTEE PRESIDENT CEO ",
" FOUNDER ",
" TRUSTEE PRESIDENT ",
" PRESIDENT CHAIR ",
" CHIEF PARTY ",
" CHAIR PRESIDENT ",
" PRESIDENT CEO TRUSTEE ",
" BOARD CHAIR ",
" PRESIDENT BOARD MEMBER ",
" BOARD MEMBER EXECUTIVE DIRECTOR ",
" FOUNDING DIRECTOR ",
" BOARD MEMBER CEO ",
" EXECUTIVE DIRECTOR BOARD MEMBER ",
" CO FOUNDER ",
" CHAIR THE BOARD ",
" CEO CHAIR ",
" FOUNDER DIRECTOR ",
" EXECUTIVE DIRECTOR TRUSTEE ",
" TRUSTEE EXECUTIVE DIRECTOR ",
" VICE CHAIR CEO ",
" CEO TRUSTEE ",
" PRESIDENTTRUSTEE ",
" BOARD MEMBER PRESIDENT CEO ",
" BOARD MEMBER PRESIDENT ",
" BOARD PRESIDENT ",
" PRESIDENT CEO BOARD MEMBER ",
" ASSOCIATE COMMISSIONER "
)


TRUST.Prob.BOARD<- c( 
" MEMBER ",
" MEMBER TEACHER ",
" SENIOR FELLOW ",
" PARTNER ",
" BOARD MEMBERDEPT CHAIR ",
" BOARD DIRECTOR ",
" FACULTY REPRESENTATIVE ",
" CHAIR DIRECTOR ",
" DIRECTOR CHAIR ",
" EX OFFICER MEMBER ",
" STAFF REPRESENTATIVE ")


TRUST.Prob.DEP.CEO <- c(" ASSOCIATE COMMISSIONER " )

TRUST.Prob.MAN <- c( 
" MEDICAL STAFF PRESIDENT ",
" STAFF REPRESENTATIVE ",
" CHAIR "
)

TRUST.Prob.OTHER<- c(
" MEMBER ",
" MEMBER TEACHER ",
" SENIOR FELLOW ",
" PARTNER ",
" EX OFFICER MEMBER ",
" FOUNDATION DIRECTOR ",
" COMMISSIONER ",
" CHIEF PARTY ",
" FOUNDING DIRECTOR ")


#' 
#' ## Human Resources
#' 
## ----Human Resources-----------------------------------------------------
HUM.RES <- c( 
" VICE PRESIDENT HUMAN RESOURCES ", 
" DIRECTOR HUMAN RESOURCES ", 
" VICE PRESIDENT HUMAN RESOURCES ", 
" VICE PRESIDENT HUMAN RESOURCES ", 
" DIRECTOR HUMAN RESOURCES ", 
" VICE PRESIDENT HUMAN RESOURCES ", 
" HUMAN RESOURCES DIRECTOR ", 
" CHIEF HUMAN RESOURCES OFFICER ", 
" SENIOR VICE PRESIDENT HUMAN RESOURCES ", 
" DIRECTOR HUMAN RESOURCES ", 
" HUMAN RESOURCES ", 
" SENIOR VICE PRESIDENT CHIEF HUMAN RESOURCES OFFICER ", 
" VICE PRESIDENT CHIEF HUMAN RESOURCES OFFICER ", 
" VICE PRESIDENT HUMAN RESOURCES ", 
" EXECUTIVE DIRECTOR HUMAN RESOURCES ", 
" SENIOR VICE PRESIDENT HUMAN RESOURCES ", 
" DIRECTOR FINANCE HUMAN RESOURCES ", 
" DIRECTOR STAFF ", 
" ASSISTANT VICE PRESIDENT HUMAN RESOURCES ",
" DIRECTOR EMPLOYEE ",
" VICE PRESIDENT PROFESSIONAL SERVICES ",
" DIRECTOR FINANCE HUMAN RESOURCES "
)

HUM.RES <- HUM.RES %>% unique


#' 
#' ## Human Resources Problems
#' 
## ----Human Resources Problems--------------------------------------------

HUM.RES.Prob <- c( " DIRECTOR STAFF " )


#' 
#' ## Communications
#' 
## ----Communications------------------------------------------------------

COMM <- c( 
" DIRECTOR MARKETING ", 
" VICE PRESIDENT MARKETING ", 
" CHIEF MARKETING OFFICER ", 
" MARKETING DIRECTOR ", 
" DIRECTOR COMMUNICATIONS ", 
" VICE PRESIDENT COMMUNICATIONS ", 
" VICE PRESIDENT MARKETING ", 
" VICE PRESIDENT EXTERNAL AFFAIRS ", 
" COMMUNICATIONS DIRECTOR ", 
" VICE PRESIDENT MARKETING COMMUNICATIONS ", 
" VICE PRESIDENT COMMUNICATIONS ", 
" VICE PRESIDENT MARKETING COMMUNICATIONS ", 
" SENIOR VICE PRESIDENT MARKETING ", 
" DIRECTOR EXTERNAL AFFAIRS ", 
" DIRECTOR COMMUNICATIONS ", 
" CHIEF COMMUNICATIONS OFFICER ", 
" DIRECTOR COMMUNICATIONS ", 
" VICE PRESIDENT MARKETING DEVELOPMENT ", 
" DIRECTOR MARKETING ", 
" VICE PRESIDENT GOVERNMENT RELATIONS ", 
" VICE PRESIDENT EXTERNAL RELATIONS ", 
" VICE PRESIDENT UNIVERSITY RELATIONS ", 
" DIRECTOR PUBLICATIONS ", 
" VICE PRESIDENT PUBLIC RELATIONS ", 
" VICE PRESIDENT COMMUNICATIONS ", 
" VICE PRESIDENT PUBLIC AFFAIRS ", 
" VICE PRESIDENT COMMUNICATIONS MARKETING ", 
" VICE PRESIDENT COMMUNICATIONS ", 
" DIRECTOR PUBLICATIONS ",
" VICE PRESIDENT GOVERNMENT AFFAIRS ",
" DIRECTOR INVESTMENTS ",
" DIRECTOR PLANNED GIVING ",
" ARTISTIC DIRECTOR ",
" ARTISTIC DIRECTOR ",
" EXECUTIVE ARTISTIC DIRECTOR ",
" PRODUCING ARTISTIC DIRECTOR ",
" ASSOCIATE ARTISTIC DIRECTOR ",
" CO ARTISTIC DIRECTOR ",
" PRESIDENT ARTISTIC DIRECTOR ",
" PROGRAM DIRECTOR ",
" ASSISTANT VICE PRESIDENT ",
" VICE PRESIDENT MARKETING DEVELOPMENT "
)

COMM <- COMM %>% unique


#' 
#' ## Management 
#' 
## ----Management----------------------------------------------------------

MAN <- c(
" DIRECTOR ",
" VICE PRESIDENT ",
" ADMINISTRATOR ",
" CONTROLLER ",
" EXECUTIVE VICE PRESIDENT ",
" SENIOR VICE PRESIDENT ",
" ASSISTANT DIRECTOR ",
" ASSOCIATE DIRECTOR ",
" MANAGER ",
" DIRECTOR ",
" OFFICER ",
" CO DIRECTOR ",
" GENERAL MANAGER ",
" DIRECTOR ",
" VICE PRESIDENT ",
" ADMINISTRATOR ",
" CONTROLLER ",
" EXECUTIVE VICE PRESIDENT ",
" SENIOR VICE PRESIDENT ",
" ASSISTANT DIRECTOR ",
" ASSOCIATE DIRECTOR ",
" MANAGER ",
" DIRECTOR ",
" OFFICER ",
" CO DIRECTOR ",
" GENERAL MANAGER ",
" FORMER OFFICER ",
" SENIOR DIRECTOR ",
" REGIONAL DIRECTOR ",
" VICE PRESIDENT DIRECTOR ",
" EX DIRECTOR ",
" COORDINATOR ",
" EX OFFICER ",
" FORMER DIRECTOR ",
" FORMER DIRECTOR ",
" DIVISION DIRECTOR ",
" DIRECTOR TEACHER ",
" ASSOCIATE VICE PRESIDENT ",
" GENERAL DIRECTOR ",
" SUPERVISOR ",
" FORMER VICE PRESIDENT ",
" EDITOR ",
" DEPARTMENT DIRECTOR ",
" DIRECTOR VICE PRESIDENT ",
" NATIONAL DIRECTOR ",
" SITE DIRECTOR ",
" STATE DIRECTOR ",
" STORE MANAGER ",
" UNIT DIRECTOR ",
" OFFICER MGR ",
" COUNTRY DIRECTOR ",
" VICE PRINCIPAL ",
" DISTRICT VICE PRESIDENT ",
" BRANCH DIRECTOR ",
" ASSIST DIRECTOR ",
" DIVISIONAL DIRECTOR ",
" GROUP VICE PRESIDENT ",
" DIVISION CHIEF ",
" FORMER EXECUTIVE VICE PRESIDENT ",
" REGIONAL VICE PRESIDENT ",
" DIRECTOR FOUNDATION ",
" FORMER DIRECTOR ",
" FORMER VICE PRESIDENT ",
" KEY EMPLOYEE ",
" FORMER KEY EMPLOYEE "
)

MAN <- MAN %>% unique

#' 
#' ## Management Problems
#' 
## ----Management Problems-------------------------------------------------

MAN.Prob <- c( 
" DEAN ", 
" DEAN ", 
" OFFICER ", 
" FORMER OFFICER ", 
" CAO ", 
" EX DIRECTOR ", 
" EX OFFICER ", 
" FORMER DIRECTOR ", 
" BUSINESS MGR ", 
" ASSOCIATE DEAN ", 
" CENTER DIRECTOR ", 
" BUSINESS MANAGEMENT ", 
" ASSISTANT PRINCIPAL ", 
" HEAD LOWER SCHOOL ", 
" GENERAL DIRECTOR ", 
" EDITOR ", 
" RESEARCH DIRECTOR ", 
" BOARD MEMBERDEPT CHAIR ", 
" REGIONAL EXECUTIVE DIRECTOR ", 
" COUNTRY DIRECTOR ", 
" HEAD COACH MEN S BASKETBALL ", 
" HEAD BASKETBALL COACH ", 
" MEN S BASKETBALL COACH " 
 )

MAN.Prob <- MAN.Prob %>% unique

#' 
#' ## Technology
#' 
## ----Technology----------------------------------------------------------

TECH <- c( 
" CHIEF INFOMATION OFFICER ", 
" DIRECTOR IT ", 
" IT DIRECTOR ", 
" VICE PRESIDENT CIO ", 
" DIRECTOR TECHNOLOGY ", 
" CHIEF TECHNOLOGY OFFICER ", 
" CDO ", 
" CIO ", 
" VICE PRESIDENT INFOMATION TECHNOLOGY ", 
" VICE PRESIDENT CHIEF INFOMATION OFFICER ", 
" SENIOR VICE PRESIDENT CIO ", 
" DIRECTOR INFOMATION TECHNOLOGY ", 
" DIRECTOR IT ", 
" CTO ", 
" IT MANAGER ", 
" VICE PRESIDENT INFOMATION SERVICES ", 
" VICE PRESIDENT INFOMATION SYSTEMS ", 
" CHIEF INFOMATION OFFICER ", 
" SENIOR VICE PRESIDENT CHIEF INFOMATION OFFICER ", 
" VICE PRESIDENT IT ", 
" DIRECTOR INFOMATION SYSTEMS ", 
" DIRECTOR INFOMATION TECHNOLOGY ", 
" DIRECTOR MANAGEMENT INFOMATION SYSTEMS ", 
" CHIEF MEDICAL INFOMATION OFFICER ", 
" VICE PRESIDENT INFOMATION TECHNOLOGY ", 
" VICE PRESIDENT TECHNOLOGY ", 
" DIRECTOR INFOMATION SYSTEMS ", 
" CHIEF INNOVATION OFFICER " ) 

TECH <- TECH %>% unique

#' 
#' 
#' ## Development
#' 
## ----Development---------------------------------------------------------

DEV <- c(
" DIRECTOR DEVELOPMENT ", 
" VICE PRESIDENT DEVELOPMENT ", 
" CHIEF DEVELOPMENT OFFICER ", 
" CHIEF DEVELOPMENT OFFICER ", 
" VICE PRESIDENT DEVELOPMENT ", 
" DIRECTOR DEVELOPMENT ", 
" VICE PRESIDENT DEVELOPMENT ", 
" CHIEF DEVELOPMENT OFFICER ", 
" CHIEF DEVELOPMENT OFFICER ", 
" VICE PRESIDENT DEVELOPMENT ", 
" VICE PRESIDENT BUSINESS DEVELOPMENT ", 
" SENIOR VICE PRESIDENT DEVELOPMENT ", 
" DIRECTOR DEVELOPMENT ", 
" VICE PRESIDENT DEVELOPMENT ", 
" DEVELOPMENT ", 
" VICE PRESIDENT INSTITUTIONAL ADVANCEMENT ", 
" DIRECTOR BUSINESS DEVELOPMENT ", 
" SENIOR DIRECTOR DEVELOPMENT ", 
" SENIOR VICE PRESIDENT BUSINESS DEVELOPMENT ", 
" DIRECTOR BUSINESS DEVELOPMENT ", 
" VICE PRESIDENT PHILANTHROPY ", 
" VICE PRESIDENT BUSINESS DEVELOPMENT ", 
" VICE PRESIDENT INSTITUTION ADVANCEMENT ", 
" DIRECTOR MAJOR GIFTS ", 
" VICE PRESIDENT FUND DEVELOPMENT ", 
" VICE PRESIDENT UNIVERSITY ADVANCEMENT ", 
" DEVELOPMENT OFFICER ", 
" DIRECTOR INSTITUTIONAL ADVANCEMENT ", 
" VICE PRESIDENT PHILANTHROPY ", 
" DO ", 
" VICE PRESIDENT RESOURCES DEVELOPMENT ", 
" VICE PRESIDENT CHIEF DEVELOPMENT OFFICER ", 
" EXECUTIVE DIRECTOR DEVELOPMENT ", 
" DIRECTOR PLANNED GIVING ", 
" SENIOR VICE PRESIDENT ADVANCEMENT ", 
" SENIOR DIRECTOR DEVELOPMENT ", 
" ASSISTANT VICE PRESIDENT DEVELOPMENT ", 
" CHIEF PHILANTHROPY OFFICER ", 
" SENIOR VICE PRESIDENT DEVELOPMENT ",
" DEVELOPMENT DIRECTOR ",
" VICE PRESIDENT INSTITUTIONAL ADVANCEMENT ",
" VICE PRESIDENT ADVANCEMENT ",
" DIRECTOR ADVANCEMENT ",
" CHIEF INVESTMENT OFFICER ",
" VICE PRESIDENT ADVANCEMENT ",
" CHIEF ADVANCEMENT OFFICER ",
" VICE PRESIDENT MARKETING DEVELOPMENT "
 ) 

DEV <- DEV %>% unique

#' 
#' ## Development Problems
#' 
## ----Development Problems------------------------------------------------

DEV.Prob <- c( " DO ") 



#' 
#' ## OTHER Revision
#' 
## ------------------------------------------------------------------------
OTHER <- c(
    " EMPLOYEE ",
" STAFF ",
" CLERK ",
" HIGHEST COMPENSATED EMPLOYEE ",
" COACH ",
" INVESTIGATOR ",
" PRINCIPAL INVESTIGATOR ",
" ASSOCIATE ",
" HIGHEST COMPENSATED EMPLOY ",
" STAGEHAND ",
" SEE SCHEDULE O ",
" SEE SCHEDULE O O T TITLES "
)

OTHER <- OTHER %>% unique()

#' 
#' ## Project Mangement
#' 
## ------------------------------------------------------------------------
PROJECT <-c(
" CHIEF PROGRAM OFFICER ",
" SENIOR VICE PRESIDENT PROGRAM ",
" DIRECTOR PROGRAM OPERATIONS ",
" VICE PRESIDENT PROGRAMMING ",
" DIRECTOR PROGRAM SERVICES ",
" PROGRAM DIRECTOR ",
" PROGRAM MANAGER ",
" PROJECT DIRECTOR ",
" VICE PRESIDENT PROGRAM ",
" DIRECTOR PROGRAM ",
" VICE PRESIDENT PROGRAM ",
" PROJECT MANAGER ",
" PROGRAM COORDINATOR ",
" SENIOR PROGRAM DIRECTOR ",
" PROGRAM OFFICER ",
" DIRECTOR PROGRAM ",
" SENIOR PROGRAM MANAGER ",
" SENIOR PROJECT MANAGER ",
" PROGRAM ADMINISTRATOR ",
" PROJECT COORDINATOR ",
" ASSISTANT PROGRAM DIRECTOR ",
" ASSISTANT PROGRAM DIRECTOR ",
" VICE PRESIDENT PROGRAM SERVICES ",
" DIRECTOR SPECIAL PROJECT ",
" DIRECTOR PROGRAMMING "
)

PROJECT <- PROJECT %>% unique()

#' 
#' ## LEGAL
#' 
## ------------------------------------------------------------------------
LEGAL <- c(
" SENIOR VICE PRESIDENT GENERAL COUNSEL ",
" CHIEF COMPLIANCE OFFICER ",
" CCO ",
" CHIEF LEGAL OFFICER ",
" LEGAL DIRECTOR ",
" COMPLIANCE OFFICER ",
" VICE PRESIDENT GENERAL COUNSEL ",
" EXECUTIVE VICE PRESIDENT GENERAL COUNSEL ",
" VICE PRESIDENT PUBLIC POLICY ",
" POLICY DIRECTOR ",
" DIRECTOR LITIGATION ",
" VICE PRESIDENT COMPLIANCE ",
" VICE PRESIDENT POLICY ",
" DIRECTOR PUBLIC POLICY ",
" DIRECTOR COMPLIANCE ",
" VICE PRESIDENT GENERAL COUNSEL ",
" VICE PRESIDENT LEGAL AFFAIRS ",
" VICE PRESIDENT GENERAL COUNSEL SECRETARY ",
" DEPUTY GENERAL COUNSEL ",
" GENERAL COUNSEL ",
" SUPERVISING ATTORNEY ",
" ATTORNEY ",
" SENIOR ATTORNEY ",
" COUNSELOR ",
" CHIEF COUNSEL ",
" LEGAL COUNSEL ",
" MEMBER COUNSEL ",
" ASSOCIATE GENERAL COUNSEL ",
" SENIOR COUNSEL ",
" ASSOCIATE GENERAL COUNSEL "
)

LEGAL <- LEGAL %>% unique()

#' 
#' 
#' ## FACILITIES
#' 
## ------------------------------------------------------------------------
FACILITIES <- c(
" DIRECTOR FACILITIES ",
" VICE PRESIDENT FACILITIES ",
" FACILITIES DIRECTOR ",
" VICE PRESIDENT FACILITIES MANAGEMENT ",
" FACILITY DIRECTOR ",
" MAINTENANCE "
)

FACILITIES <- FACILITIES %>% unique()

#' 
#' ## Administrative Support
#' 
## ------------------------------------------------------------------------
ADMIN.SUP <- c(
" VICE PRESIDENT SUPPORT SERVICES ",
" VICE PRESIDENT ANCILLARY SERVICES ",
" VICE PRESIDENT ADMINISTRATIVE SERVICES ",
" DIRECTOR SUPPORT SERVICES ",
" ASSISTANT ADMINISTRATOR ",
" OFFICER ADMINISTRATION ",
" ASSISTANT SE ",
" EXECUTIVE ASSISTANT ",
" ADMINISTRATION ASSISTANT ",
" ADMINISTRATIVE ASSISTANT ",
" OFFICER MANAGEMENT "
)

ADMIN.SUP <- ADMIN.SUP %>% unique()

#' 
#' ## Health and Medical
#' 
## ------------------------------------------------------------------------
MED.MAN <- c(
" VICE PRESIDENT CNE ", 
" MEDICAL DIRECTOR ",
" CHIEF MEDICAL OFFICER ",
" DIRECTOR NURSING ",
" CHIEF NURSING OFFICER ",
" CNO ",
" VICE PRESIDENT MEDICAL AFFAIRS ",
" DIRECTOR PHARMACY ",
" PHARMACY DIRECTOR ",
" ASSOCIATE MEDICAL DIRECTOR ",
" DIRECTOR PHARMACY ",
" VICE PRESIDENT PATIENT CARE SERVICES ",
" PHARMACY MANAGER ",
" VICE PRESIDENT CLINICAL SERVICES ",
" CHIEF CLINICAL OFFICER ",
" MED DIRECTOR ",
" VICE PRESIDENT PATIENT CARE ",
" VICE PRESIDENT CHIEF NURSING OFFICER ",
" ASSISTANT MEDICAL DIRECTOR ",
" CLINIC DIRECTOR ",
" SENIOR VICE PRESIDENT MEDICAL AFFAIRS ",
" VICE PRESIDENT PATIENT SERVICES ",
" SENIOR VICE PRESIDENT CHIEF NURSING OFFICER ",
" VICE PRESIDENT NURSING ",
" NURSING SUPERVISOR ",
" PHYSICIAN DIRECTOR ",
" NURSE MANAGER ",
" NURSING DIRECTOR ",
" DIRECTOR NURSING ",
" VICE PRESIDENT MEDICAL AFFAIRS ",
" DIRECTOR CLINICAL SERVICES ",
" SENIOR VICE PRESIDENT CNO ",
" VICE PRESIDENT BUSINESS ",
" CLINICAL MANAGER ",
" SENIOR VICE PRESIDENT CHIEF MEDICAL OFFICER ",
" VICE PRESIDENT CHIEF MEDICAL OFFICER ",
" VICE PRESIDENT MED AFFAIRS ",
" DIRECTOR REHAB ",
" VICE PRESIDENT CLINICAL ",
" CHIEF DENTAL OFFICER ",
" CHIEF MEDICAL DIRECTOR ",
" DIRECTOR HEALTH SERVICES ",
" ASSISTANT DIRECTOR NURSING ",
" PHYSICIAN MEDICAL DIRECTOR ",
" VICE PRESIDENT MANAGED CARE ",
" MANAGER PHARMACY ",
" VICE PRESIDENT AMBULATORY SERVICES ",
" VICE PRESIDENT PHYSICIAN SERVICES ",
" VICE PRESIDENT NURSING SERVICES ",
" CLINICAL COORDINATOR ",
" DIRECTOR REHABILITATION ",
" VICE PRESIDENT CLINICAL OPERATIONS ",
" CHIEF NURSE EXECUTIVE ",
" VICE PRESIDENT MEDICAL AFFAIRS CMO ",
" REGIONAL MEDICAL DIRECTOR ",
" VICE PRESIDENT SURGICAL SERVICES ",
" DEAN SCHOOL NURSING ",
" EXECUTIVE MEDICAL DIRECTOR ",
" MDS COORDINATOR ",
" VICE PRESIDENT PATIENT CARE SERVICE ",
" ASSISTANT VICE PRESIDENT NURSING ",
" CHIEF MED OFFICER ",
" DIRECTOR CARE AIDE ",
" SENIOR MEDICAL DIRECTOR ",
" VICE PRESIDENT NURSING ",
" DIRECTOR CLINICAL SERVICES ",
" DIRECTOR HEALTH SERVICES ",
" DIRECTOR REHAB ",
" DIRECTOR CLINICAL SERVICE ",
" NURSING HOME ADMINISTRATOR ",
" VICE PRESIDENT HEALTH SERVICES ",
" DIRECTOR SURGICAL SERVICES ",
" ASSISTANT DIRECTOR NURSING ",
" DIRECTOR PATIENT SERVICES ",
" DEAN SCHOOL NURSING ",
" VICE PRESIDENT MEDICAL SERVICES ",
" DIRECTOR PHARMACY SERVICES ",
" SENIOR VICE PRESIDENT MANAGED CARE ",
" SENIOR VICE PRESIDENT PATIENT CARE SERVICES ",
" DIRECTOR REHAB SERVICES ",
" VICE PRESIDENT PATIENT CARE SERVICES CNO ",
" DIRECTOR DINING SERVICES ",
" VICE PRESIDENT NURSING CNO ",
" CHIEF MEDICAL ",
" VICE PRESIDENT CNE ",
" NURSING HOME ADMINISTRATOR ",
" VICE PRESIDENT CLINICAL OPERATIONS ",
" CLINICAL DIRECTOR ",
" DENTAL DIRECTOR ",
" VICE PRESIDENT CNO ",
" DIRECTOR PHYSICIAN "
)

MED.MAN <- MED.MAN %>% unique()


#' 
#' 
#' ## Health and Human Services
#' 
## ------------------------------------------------------------------------
HEALTH.HUM <- c(
" SHELTER MANAGER ",
" HOUSING DIRECTOR ",
" SHELTER DIRECTOR ",
" CASE MANAGER "
)

HEALTH.HUM <- HEALTH.HUM %>% unique()

#' 
#' 
#' ## Training and Education
#' 
#' 
## ------------------------------------------------------------------------
TRAIN <- c( 
" DIRECTOR EDUCATION ",
" TRAINING DIRECTOR ",
" EDUCATION DIRECTOR ",
" VICE PRESIDENT EDUCATION ",
" TRAINING COORDINATOR ",
" DIRECTOR TRAINING ",
" VICE PRESIDENT EDUCATION ",
" EDUCATIONAL DIRECTOR ",
" CHIEF LEARNING OFFICER ",
" DIRECTOR INSTRUCTION ",
" DIRECTOR TRAINING ",
" DIRECTOR INSTRUCTION ",
" VICE PRESIDENT CNE "
)

TRAIN <- TRAIN %>% unique()


#' 
#' ## Academic Management
#' 
## ------------------------------------------------------------------------
ACADEMIC.MAN <- c( 
    " VICE PRESIDENT ACADEMIC AFFAIRS ",
" VICE PRESIDENT STUDENT AFFAIRS ",
" CHIEF ACADEMIC OFFICER ",
" DEAN FACULTY ",
" VICE PRESIDENT ACADEMIC AFFAIRS ",
" CHIEF ACADEMICS ",
" DEAN SCHOOL BUSINESS ",
" VICE PRESIDENT STUDENT LIFE ",
" VICE PRESIDENT STUDENT DEVELOPMENT ",
" VICE PRESIDENT ACADEMICS ",
" VICE PRESIDENT STUDENT AFFAIRS ",
" VICE PRESIDENT STUDENT LIFE ",
" DEAN SCHOOL EDUCATION ",
" VICE PRESIDENT ACADEMIC ",
" DEAN ADMISSIONS ",
" DIRECTOR COLLEGE COUNSELING ",
" ACADEMIC DIRECTOR ",
" VICE PRESIDENT STUDENT SERVICES ",
" DEAN LAW SCHOOL ",
" VICE PRESIDENT STUDENT DEVELOPMENT ",
" DEAN ACADEMIC AFFAIRS ",
" DEAN COLLEGE BUSINESS ",
" VICE PRESIDENT ACADEMIC AFFAIRS ",
" DEAN STUDENTS ",
" CHIEF ACADEMIC OFFICER ",
" ACADEMIC DEAN ",
" DEAN FACULTY ",
" CHIEF ACADEMICS ",
" DEAN SCHOOL BUSINESS ",
" PROVOST VICE PRESIDENT ACADEMIC AFFAIRS ",
" DIRECTOR COLLEGE COUNSELING ",
" VICE PRESIDENT COLLEGE ADVANCEMENT ",
" DEAN SCHOOL EDUCATION ",
" VICE PRESIDENT ACADEMIC ",
" DEAN ADMISSIONS ",
" ACADEMIC DIRECTOR ",
" DIRECTOR SURGERY ",
" DEAN ACADEMIC AFFAIRS ",
" DEAN COLLEGE BUSINESS ",
" VICE PRESIDENT DEAN ",
" DEAN THE COLLEGE ",
" HEAD UPPER SCHOOL ",
" DEAN LAW SCHOOL ",
" ASSOCIATE PROFESSOR ",
" ASSOCIATE DEAN ",
" ACADEMIC DEAN ",
" ASSOCIATE PROVOST ",
" VICE CHANCELLOR ",
" DEPARTMENT CHAIR ",
" DEAN "
)

ACADEMIC.MAN <- ACADEMIC.MAN %>% unique()

#' 
#' 
#' ## Department Head
#' 
## ------------------------------------------------------------------------
DEP.HEAD <- c(
" DIRECTOR ADMISSIONS ",
" ATHLETIC DIRECTOR ",
" LIBRARY DIRECTOR ",
" VICE PRESIDENT ENROLLMENT MANAGEMENT ",
" MUSIC DIRECTOR ",
" CHIEF STRATEGY OFFICER ",
" CPO ",
" DIRECTOR RESEARCH ",
" DIRECTOR ATHLETICS ",
" HEAD MIDDLE SCHOOL ",
" VICE PRESIDENT RESEARCH ",
" VICE PRESIDENT ENROLLMENT ",
" DIRECTOR SALES ",
" BUSINESS DIRECTOR ",
" VICE PRESIDENT ENROLLMENT MANAGEMENT ",
" TECHNICAL DIRECTOR ",
" CHIEF PROFESSIONAL OFFICER ",
" VICE PRESIDENT ENROLLMENT ",
" DIRECTOR COACHING ",
" CHIEF QUALITY OFFICER ",
" VICE PRESIDENT QUALITY ",
" EXECUTIVE VICE PRESIDENT CAO ",
" SENIOR VICE PRESIDENT STRATEGY PLANNING ",
" CHIEF CURATOR ",
" DIRECTOR R D ",
" VICE PRESIDENT PROFESSIONAL SERVICES ",
" HEAD FOOTBALL COACH ",
" SENIOR VICE PRESIDENT CAO ",
" HEAD START DIRECTOR ",
" CHIEF SCIENTIFIC OFFICER ",
" VICE PRESIDENT ADMISSIONS ",
" VICE PRESIDENT ENROLLMENT ",
" VICE PRESIDENT STRATEGY PLANNING ",
" CHIEF BUSINESS OFFICER ",
" DIRECTOR INVESTMENTS ",
" VICE PRESIDENT BUSINESS AFFAIRS ",
" SALES MANAGER ",
" SENIOR VICE PRESIDENT PLANNING ",
" LAB DIRECTOR ",
" PRODUCING DIRECTOR ",
" VICE PRESIDENT SALES ",
" SENIOR VICE PRESIDENT MISSION INTEGRATION ",
" EXECUTIVE DIRECTOR LEADERSHIP INSTITUTION ",
" DIRECTOR PRODUCTION ",
" CREATIVE DIRECTOR ",
" VICE PRESIDENT MISSION ",
" DIRECTOR ADMISSION ",
" ADMISSIONS DIRECTOR ",
" VICE PRESIDENT PLANNING ",
" TEACHER DIRECTOR ",
" VICE PRESIDENT ENGINEERING ",
" VICE PRESIDENT MISSION INTEGRATION ",
" CBO ",
" DIRECTOR RESEARCH ",
" SCIENTIFIC DIRECTOR ",
" SENIOR VICE PRESIDENT CHIEF QUALITY OFFICER ",
" DIRECTOR ADMISSIONS ",
" DIRECTOR ADVANCEMENT ",
" DIRECTOR ATHLETICS ",
" DIRECTOR SALES ",
" DIRECTOR FINANCIAL SERVICES ",
" DEAN ARTS SCIENCES ",
" VICE PRESIDENT FINANCIAL AFFAIRS ",
" VICE PRESIDENT STRATEGY ",
" VICE PRESIDENT BUSINESS SERVICES ",
" DIRECTOR ADMISSION ",
" DIRECTOR FINANCIAL SERVICE ",
" DIRECTOR PRODUCTION ",
" DIRECTOR RISK MANAGEMENT ",
" VICE PRESIDENT ENROLLMENT SERVICES ",
" VICE PRESIDENT RISK MANAGEMENT ",
" VICE PRESIDENT RETAIL ",
" DIRECTOR MEMBERSHIP ",
" CHIEF PROFESSIONAL OFFICER ",
" CHIEF CURATOR ",
" HEAD FOOTBALL COACH ",
" EXECUTIVE DIRECTOR LEADERSHIP INSTITUTION ",
" BUSINESS MANAGER ",
" BUSINESS MANAGER ",
" CAO ",
" BUSINESS MGR ",
" CENTER DIRECTOR ",
" BUSINESS MANAGEMENT ",
" ASSISTANT PRINCIPAL ",
" HEAD LOWER SCHOOL ",
" RESEARCH DIRECTOR ",
" VICE PRESIDENT FOUNDATION ",
" VICE PRESIDENT ENROLLMENT MANAGEMENT ",
" BUSINESS ADMINISTRATOR ",
" HEAD COACH MEN S BASKETBALL ",
" HEAD BASKETBALL COACH ",
" MEN S BASKETBALL COACH ",
" BUSINESS OFFICER "
)

DEP.HEAD <- DEP.HEAD %>% unique()

#' 
#' ## Profesional
#' 
## ------------------------------------------------------------------------
PROFESIONAL <- c(
" PHYSICIAN ",
" PSYCHIATRIST ",
" PHARMACIST ",
" PROFESSOR ",
" DENTIST ",
" CRNA ",
" DOCTOR ",
" MD ",
" TEACHER ",
" FACULTY ",
" KEY EMPLOYEE ",
" STAFF PHYSICIAN ",
" RN ",
" REGISTERED NURSE ",
" NURSE PRACTITIONER ",
" PHYSICIAN ",
" PSYCHIATRIST ",
" PHARMACIST ",
" PROFESSOR ",
" DENTIST ",
" CRNA ",
" DOCTOR ",
" MD ",
" TEACHER ",
" FACULTY ",
" STAFF PHYSICIAN ",
" RN ",
" REGISTERED NURSE ",
" NURSE PRACTITIONER ",
" SURGEON ",
" PHYSICAL THERAPIST ",
" MEDICAL DOCTOR ",
" PHYSICIAN ASSISTANT ",
" INSTRUCTOR ",
" CARDIOLOGIST ",
" ANESTHESIOLOGIST ",
" PEDIATRICIAN ",
" PROVIDER ",
" ER PHYSICIAN ",
" NURSE ",
" PASTOR ",
" ORTHOPEDIC SURGEON ",
" HOSPITALIST ",
" SENIOR SCIENTIST ",
" OB GYN ",
" RADIOLOGIST ",
" STAFF PSYCHIATRIST ",
" PATHOLOGIST ",
" NEUROSURGEON ",
" ASSISTANT PROFESSOR ",
" OB GYN PHYSICIAN ",
" GENERAL SURGEON ",
" PHYSICAN ",
" SENIOR ASSOCIATE ",
" MUSICIAN ",
" INTERNIST ",
" CLINICAL PHARMACIST ",
" EMPLOYED PHYSICIAN ",
" NURSE ANESTHETIST ",
" PROFESSOR LAW ",
" FAMILY PRACTICE PHYSICIAN ",
" PSYCHOLOGIST ",
" PHYSICIST ",
" M D ",
" PARTNER ",
" SCIENTIST ",
" THERAPIST ",
" LPN ",
" CURATOR ",
" MINISTER ",
" NURSE PRACTITIONER ",
" VETERINARIAN ",
" CONSULTANT ",
" HIGHEST PAID ",
" MEDICAL PHYSICIST ",
" SENIOR ADVISOR ",
" FAMILY PHYSICIAN ",
" ADVANCED NURSE PRACTITIONER ",
" PUBLISHER ",
" RESEARCH SCIENTIST ",
" HOSPICE PHYSICIAN ",
" PEDIATRIC PHYSICIAN ",
" ATTENDING PHYSICIAN ",
" MEDICAL PROVIDER ",
" LAW PROFESSOR ",
" REGISTRAR ",
" ORTHOPAEDIC SURGEON ",
" RESEARCHER ",
" SENIOR RESEARCH SCIENTIST ",
" ONCOLOGIST ",
" PA ",
" PRIMARY CARE PHYSICIAN ",
" PHYSICIANS ASSISTANT ",
" STAFF PHARMACIST ",
" LIBRARIAN ",
" UROLOGIST ",
" FAMILY PRACTITIONER ",
" STAFF NURSE ",
" CLINICIAN ",
" FACULTY PHYSICIAN ",
" ADMINISTRATIVE ",
" NEUROLOGIST ",
" PHYSICIAN OB GYN ",
" PSYCHIATRIC NURSE ",
" SPEECH THERAPIST ",
" MISSIONARY ",
" GASTROENTEROLOGIST ",
" PRINCIPAL SCIENTIST ",
" TEACHER EX OFFICER SCHOOL ",
" CHARGE NURSE ",
" BASKETBALL COACH ",
" EMERGENCY PHYSICIAN ",
" RADIATION ONCOLOGIST ",
" OCCUPATIONAL THERAPIST ",
" CHIEF PHYSICIST ",
" RN SUPERVISOR ",
" APRN ",
" CHILD PSYCHIATRIST ",
" ANESTHETIST ",
" ASSISTANT FEDERAL DEFENDER ",
" LEAD TEACHER ",
" MEMBER COUNSEL ",
" OPHTHALMOLOGIST ",
" SENIOR RESEARCH FELLOW ",
" PHYSICAL THE ",
" PHYSICIAN S ASSISTANT ",
" ASSISTANT CLERK ",
" FNP ",
" ACCOUNT EXECUTIVE ",
" EDUCATOR ",
" INTERNAL MEDICINE ",
" MEDICAL STAFF ",
" PHYSICIAN CARDIOLOGY ",
" HIGHEST COMPENSATED ",
" PARAMEDIC ",
" CAPTAIN ",
" CARDIOVASCULAR SURGEON ",
" DDS ",
" NEONATOLOGIST ",
" TRAINING ",
" ASSISTANT ",
" PA C ",
" PHYSIATRIST ",
" PHYSICIAN EMPLOYEE ",
" PROFESSOR LAW ",
" FINANCIAL ",
" FAMILY PRACTICE ",
" HIGHEST PAID EMPLOYEE ",
" FORMER KEY EMPLOYEE "
)

PROFESIONAL <- PROFESIONAL %>% unique()


#' 
#' ## Other Professional
#' 
## ------------------------------------------------------------------------
OTHER.PROF <- c(
" INSTRUCTOR ",
" PASTOR ",
" SENIOR ASSOCIATE ",
" PARTNER ",
" CURATOR ",
" MINISTER ",
" SCIENTIST ",
" THERAPIST ",
" LPN ",
" NURSE PRACTITIONER ",
" VETERINARIAN ",
" CONSULTANT ",
" HIGHEST PAID ",
" SENIOR ADVISOR ",
" PUBLISHER ",
" RESEARCH SCIENTIST ",
" REGISTRAR ",
" RESEARCHER ",
" SENIOR RESEARCH SCIENTIST ",
" LIBRARIAN ",
" ADMINISTRATIVE ",
" SPEECH THERAPIST ",
" MISSIONARY ",
" PRINCIPAL SCIENTIST ",
" TEACHER EX OFFICER SCHOOL ",
" BASKETBALL COACH ",
" ASSISTANT FEDERAL DEFENDER ",
" SENIOR RESEARCH FELLOW ",
" LEAD TEACHER ",
" ASSISTANT CLERK ",
" EDUCATOR ",
" ACCOUNT EXECUTIVE ",
" HIGHEST COMPENSATED ",
" CAPTAIN ",
" TRAINING ",
" ASSISTANT ",
" PA C ",
" HIGHEST PAID EMPLOYEE ",
" FORMER KEY EMPLOYEE ",
" FINANCIAL "
)

OTHER.PROF <- OTHER.PROF %>% unique()

#' 
#' ## Medical Professional
#' 
## ------------------------------------------------------------------------

MED.PROF <- c(
" PHYSICIAN ",
" PSYCHIATRIST ",
" PHARMACIST ",
" DENTIST ",
" CRNA ",
" DOCTOR ",
" MD ",
" STAFF PHYSICIAN ",
" RN ",
" REGISTERED NURSE ",
" NURSE PRACTITIONER ",
" PHYSICIAN ",
" PSYCHIATRIST ",
" PHARMACIST ",
" DENTIST ",
" CRNA ",
" DOCTOR ",
" MD ",
" STAFF PHYSICIAN ",
" RN ",
" REGISTERED NURSE ",
" NURSE PRACTITIONER ",
" SURGEON ",
" PHYSICAL THERAPIST ",
" MEDICAL DOCTOR ",
" PHYSICIAN ASSISTANT ",
" CARDIOLOGIST ",
" ANESTHESIOLOGIST ",
" PEDIATRICIAN ",
" PROVIDER ",
" ER PHYSICIAN ",
" NURSE ",
" ORTHOPEDIC SURGEON ",
" HOSPITALIST ",
" SENIOR SCIENTIST ",
" OB GYN ",
" RADIOLOGIST ",
" STAFF PSYCHIATRIST ",
" PATHOLOGIST ",
" NEUROSURGEON ",
" ASSISTANT PROFESSOR ",
" OB GYN PHYSICIAN ",
" GENERAL SURGEON ",
" PHYSICAN ",
" MUSICIAN ",
" INTERNIST ",
" CLINICAL PHARMACIST ",
" EMPLOYED PHYSICIAN ",
" NURSE ANESTHETIST ",
" PROFESSOR LAW ",
" FAMILY PRACTICE PHYSICIAN ",
" PSYCHOLOGIST ",
" PHYSICIST ",
" M D ",
" MEDICAL PHYSICIST ",
" FAMILY PHYSICIAN ",
" ADVANCED NURSE PRACTITIONER ",
" HOSPICE PHYSICIAN ",
" PEDIATRIC PHYSICIAN ",
" ATTENDING PHYSICIAN ",
" MEDICAL PROVIDER ",
" ORTHOPAEDIC SURGEON ",
" ONCOLOGIST ",
" PA ",
" PRIMARY CARE PHYSICIAN ",
" PHYSICIANS ASSISTANT ",
" STAFF PHARMACIST ",
" UROLOGIST ",
" FAMILY PRACTITIONER ",
" STAFF NURSE ",
" CLINICIAN ",
" FACULTY PHYSICIAN ",
" NEUROLOGIST ",
" PHYSICIAN OB GYN ",
" PSYCHIATRIC NURSE ",
" GASTROENTEROLOGIST ",
" CHARGE NURSE ",
" EMERGENCY PHYSICIAN ",
" RADIATION ONCOLOGIST ",
" OCCUPATIONAL THERAPIST ",
" CHIEF PHYSICIST ",
" RN SUPERVISOR ",
" APRN ",
" CHILD PSYCHIATRIST ",
" ANESTHETIST ",
" OPHTHALMOLOGIST ",
" PHYSICAL THE ",
" PHYSICIAN S ASSISTANT ",
" FNP ",
" INTERNAL MEDICINE ",
" MEDICAL STAFF ",
" PHYSICIAN CARDIOLOGY ",
" PARAMEDIC ",
" CARDIOVASCULAR SURGEON ",
" DDS ",
" NEONATOLOGIST ",
" PHYSIATRIST ",
" PHYSICIAN EMPLOYEE ",
" FAMILY PRACTICE "
)
MED.PROF <- MED.PROF %>% unique()

#' 
#' ## Academic Professional
#' 
## ------------------------------------------------------------------------
ACADEMIC.PROF <- c(
" PROFESSOR ",
" TEACHER ",
" FACULTY ",
" PROFESSOR ",
" TEACHER ",
" FACULTY ",
" LAW PROFESSOR ",
" PROFESSOR LAW "
)
ACADEMIC.PROF <- ACADEMIC.PROF %>% unique()


#' 
#' 1. Seperate Problem Titles From Clean Titles
#' 2. Create Role Variables
#' 3. Assign Role to Clean Titles
#' 4. Use Binary Coding on Problem Titles
#' 5. Classify Unknown Titles
#' 6. Code random 1000 Non Appearing Titles
#' 7. Machine Learning Code
#' 
#' # Creates list all Titles
## ------------------------------------------------------------------------

Titles <- list (CEO = CEO, CFO = CFO, Treasurer = Treasurer, DEP.CEO = DEP.CEO, SEC = SEC, COO = COO, TRUST = TRUST, COMM = COMM, HUM.RES = HUM.RES, DEP.HEAD = DEP.HEAD, MAN = MAN, TECH = TECH, DEV = DEV, OTHER = OTHER, PROJECT = PROJECT, LEGAL = LEGAL, FACILITIES = FACILITIES, ADMIN.SUP = ADMIN.SUP, MED.MAN = MED.MAN, HEALTH.HUM = HEALTH.HUM, TRAIN = TRAIN, ACADEMIC.MAN = ACADEMIC.MAN, PROFESIONAL = PROFESIONAL, OTHER.PROF = OTHER.PROF, ACADEMIC.PROF = ACADEMIC.PROF, MED.PROF = MED.PROF)


#' # Creates List Clear Titles
#' 
## ------------------------------------------------------------------------
CEO.Clear <- CEO[!(CEO %in% CEO.Prob)]
CFO.Clear <- CFO[!(CFO %in% CFO.Prob)]
Treasurer.Clear <- Treasurer[!(Treasurer %in% Treasurer.Prob)]
DEP.CEO.Clear <- DEP.CEO[!(DEP.CEO %in% DEP.CEO.Prob)]
SEC.Clear <- SEC[!(SEC %in% SEC.Prob)]
COO.Clear <- COO[!(COO %in% COO.Prob)]
TRUST.Clear <- TRUST[!(TRUST %in% TRUST.Prob)]
HUM.RES.Clear <- HUM.RES[!(HUM.RES %in% HUM.RES.Prob)]
MAN.Clear <- MAN[!(MAN %in% MAN.Prob)]
DEV.Clear <- DEV[!(DEV %in% DEV.Prob)]
COMM.Clear <- COMM
TECH.Clear <- TECH

Clear.Titles <- list (CEO.Clear = CEO.Clear, CFO.Clear = CFO.Clear, Treasurer.Clear = Treasurer.Clear, DEP.CEO.Clear = DEP.CEO.Clear, SEC.Clear = SEC.Clear, COO.Clear = COO.Clear, TRUST.Clear = TRUST.Clear, HUM.RES.Clear = HUM.RES.Clear, COMM.Clear = COMM.Clear, MAN.Clear = MAN.Clear, TECH.Clear = TECH.Clear, DEV.Clear = DEV.Clear)


#' 
#' # Creates List Problem Titles
#' 
## ------------------------------------------------------------------------
Problem.Titles <- list( CEO.Board = CEO.Board, CEO.Prob = CEO.Prob, CFO.Prob = CFO.Prob, Treasurer.Prob = Treasurer.Prob, DEP.CEO.Prob = DEP.CEO.Prob, SEC.Prob = SEC.Prob, COO.Prob = COO.Prob, TRUST.Prob = TRUST.Prob, HUM.RES.Prob = HUM.RES.Prob, MAN.Prob = MAN.Prob, DEV.Prob = DEV.Prob, CEO.SCHOOL = CEO.SCHOOL, CEO.Prob.Other = CEO.Prob.Other, TRUST.Prob.BOARD = TRUST.Prob.BOARD, TRUST.Prob.DEP.CEO = TRUST.Prob.DEP.CEO, TRUST.Prob.MAN = TRUST.Prob.MAN, TRUST.Prob.OTHER = TRUST.Prob.OTHER, SCHOOL.HEAD = SCHOOL.HEAD )

#' 
#' 
#' # CreatesSaves Lists all titles
#' 
## ------------------------------------------------------------------------
All.Titles <- c(Titles, Clear.Titles, Problem.Titles)

save( All.Titles, file = "title-categories/title-list.rdata")
