==> old-R/step-02-categorize-titles.R <==
#' ---
#' title: "Untitled"
#' author: "David Selby"
#' date: "7/9/2019"
#' output: html_document
#' ---
#' 
## ----setup, include=FALSE------------------------------------------------
knitr::opts_chunk$set( echo = TRUE, message=F, warning=F, cache = T )

#' 
#' # Required Packages
#' 
## ------------------------------------------------------------------------
require( dplyr )
require( pander )
require( stringr )
require( hunspell )

#' 
#' # Load Data
#' 
#' 
## ------------------------------------------------------------------------
d2 <- readRDS( "cleaned-titles/cleaned-titles-2016.rds")


#' 
#' 
#' ## Title Catagories  
#' 
#' Shows the sorting of Standardized and Cleaned into titles into Catagories in the following steps:
#' 
#' 1. Individuals are sorted to Catagories based on if they have one the 1000 most common titles.   
#' 2. The most common words or phrases associated with the catagories are used to sort remaining individuals into catagories (where they exist)  
#' 3. Catagorized Titles are used to create meta-titles (C-Suite, Management, Board, ect)  
#' 
#' 
#' 
#' ### Catagories
#' 
#' Variable Name	        Category Name               
#' --------------       --------------			          
#' CEO	                  Chief Executive		          
#' CFO	                  Financial				            
#' Treasurer	            Treasurer				            
#' DEP.CEO	              Deputy Chief Executive		  
#' SEC	                  Secretary				            
#' COO	                  Operations				          
#' TRUST	                Trustee/Board				        
#' HUM.RES	              Human Resources				      
#' DEP.HEAD	            Department Head				      
#' MAN	Generic           Management				          
#' DEV	                  Development				          
#' TECH	                Technology/IT				        
#' COMM	                Communications				      
#' OTHER	                Other Category				      
#' PROJECT	              Project Manager				      
#' LEGAL	                Legal				           		  
#' FACILITIES            Facilities				          
#' ADMIN.SUP	            Administrative Support		  
#' MED.MAN	              Medical Management				  
#' HEALTH.HUM	          Health and Human Services		
#' TRAIN	                Training/Education		      
#' ACADEMIC.MAN	        Academic Manager				    
#' PROFESIONAL	          Professional				        
#' OTHER.PROF	          Other Professional				  
#' ACADEMIC.PROF	        Academic Professional		    
#' MED.PROF	            Medical Professional			  
#' 
#' 
#' ### Form 990 Title Instructions
#' 
#' **Titles**
#' 
#' * Current officers, directors, and trustees ( no minimum compensation threshold ).
#' 
#' * Current key employees ( over $150,000 of reportable compensation ).
#' 
#' * Current five highest compensated employees other than officers, directors, trustees, or listed key employees ( over $100,000 of reportable compensation ).
#' 
#' * Former officers, key employees, and highest compensated employees ( over $100,000 of reportable compensation, with special rules for former highest compensated employees ).
#' 
#' * Former directors and trustees ( over $10,000 of reportable compensation in the capacity as a former director or trustee ).
#' 
#' **Director or Trustee**
#' 
#' A “director or trustee” is a member of the organization's governing body, but only if the member has voting rights. A director or trustee that served at any time during
#' the organization's tax year is deemed a current director or trustee. Members of advisory boards that don't exercise any governance authority over the organization aren't considered
#' directors or trustees.
#' 
#' An “institutional trustee” is a trustee that isn't an individual or natural person but an organization. For instance, a bank or trust company serving as the trustee of a trust is an institutional trustee.
#' 
#' **Officer**
#' 
#' An officer is a person elected or appointed to manage the organization's daily operations. An officer that served at any time during the organization's tax year is deemed a current officer. The officers of an organization are determined by reference to its organizing document, bylaws, or resolutions of its governing body, or as otherwise designated consistent with state law, but, at a minimum, include those officers required by applicable state law. Officers can include a president, vice-president, secretary, treasurer and, in some cases, a Board Chair. In addition, for purposes of Form 990, including Part VII, Section A, and Schedule J ( Form 990 ), treat as an officer the following persons, regardless of their titles.
#' 
#' 1. *Top management official*. The person who has ultimate responsibility for implementing the decisions of the governing body or for supervising the management, administration, or operation of the organization; for example, the organization's president, CEO, or executive director.
#' 
#' 2. *Top financial official*. The person who has ultimate responsibility for managing the organization's finances; for example, the organization's treasurer or chief financial officer.
#' 
#' If ultimate responsibility resides with two or more individuals ( for example, co-presidents or co-treasurers ), who can exercise such responsibility in concert or individually, then treat all such individuals as officers.
#' 
#' **Key Employees**
#' 
#' Key employee. For purposes of Form 990, a current key employee is an employee of the organization ( other than an officer, director, or trustee ) who meets all three of the following tests, applied in the following order:
#' 
#' 1. \$150,000 Test: Receives reportable compensation from the organization and all related organizations in excess of \$150,000 for the calendar year ending with or within the organization's tax year.
#' 
#' 2. Responsibility Test: At any time during the calendar year ending with or within the organization's tax year:
#' 
#'   a. Has responsibilities, powers, or influence over the organization as a whole that is similar to those of officers, directors, or trustees;
#' 
#'   b. Manages a discrete segment or activity of the organization that represents 10% or more of the activities, assets, income, or expenses of the organization, as compared to the organization as a whole; or
#' 
#'   c. Has or shares authority to control or determine 10% or more of the organization's capital expenditures, operating budget, or compensation for employees.
#' 
#' 3. Top 20 Test: Is one of the 20 employees other than officers, directors, and trustees who satisfy the \$150,000 Test and Responsibility Test with the highest reportable compensation from the organization and related organizations for the calendar year ending with or within the organization's tax year.
#' 
#'   If the organization has more than 20 individuals who meet the \$150,000 Test and Responsibility Test, report as key employees only the 20 individuals that have the highest reportable compensation from the organization and related organizations. Note that any others, up to five, might be reportable as current highest compensated employees, with over $100,000 in reportable compensation. Use the calendar year ending with or within the organization's tax year for determining the organization's current key employees.
#'   
#'   An individual that isn't an employee of the organization ( or of a disregarded entity of the organization ) is nonetheless treated as a key employee if he or she serves as an officer or director of a disregarded entity of the organization and otherwise meets the standards of a key employee set forth above. See Disregarded Entities, later, for treatment of certain employees of a disregarded entity as a key employee of the organization.
#'   
#'   If an employee is a key employee of the organization for only a portion of the year, that person's entire compensation for the calendar year ending with or within the organization's tax year, from both the filing organization and related organizations,should be reported in Part VII, Section A.
#'   
#'   Management companies and similar entities that are independent contractors should not be reported as key employees. The organization's top management official and top financial official are deemed officers rather than key employees.
#'   
#' In the examples set forth below, assume the individual involved is an employee that satisfies the \$150,000 Test and Top 20 Test and isn't an officer, director, or trustee.\\
#'   
#'   **Example 1.** T is a large section 501( c )( 3 ) university. L is the dean of the law school of T, which generates more than 10% of the revenue of T, including contributions from alumni and foundations. Although L does not have ultimate responsibility for managing the university as a whole, L meets the Responsibility Test and is reportable as a key employee of T.
#'   
#'   **Example 2.** S chairs a small academic department in the College of Arts and Sciences of the same university, T, described above. As department chair, S supervises faculty in the department, approves the course curriculum, and oversees the operating budget for the department. The department represents less than 10% of the university's activities, assets,
#' income, expenses, capital expenditures, operating budget, and employee compensation. Under these facts and circumstances, S does not meet the Responsibility Test and isn't a key
#' employee of T.
#' 
#'   **Example 3.** U is a large acute-care section 501( c )( 3 ) hospital. U employs X as a radiologist. X gives instructions to staff for the radiology work X conducts, but X does not supervise other U employees, manage the radiology department, or have or share authority to control or determine 10% or more of U's capital expenditures, operating budget, or employee compensation. Under these facts and circumstances, X does not meet the Responsibility Test and isn't a key employee of U. \\
#' 
#'   **Example 4.** W is a cardiologist and head of the cardiology department of the same hospital U described above. The cardiology department is a major source of patients admitted to U and consequently represents more than 10% of U's income, as compared to U as a whole. As department head, W manages the cardiology department. Under these facts and circumstances, W meets the Responsibility Test and is a key employee of U. \\
#' 
#' 
#' 
#' 
#' 
#' 
#' ### Load Categorized Titles
#' 
#' This file indicates
#' 
#' 
## ------------------------------------------------------------------------
Title.Categorized <- load("title-categories/title-list.rdata")
All.Titles %>% names()

#' 
#' Creates new Title Categories - These are NOT mutually exclusive.
#' 
## ------------------------------------------------------------------------
d3 <- d2


d2$CEO           <- 0
d2$CFO           <- 0
d2$Treasurer     <- 0
d2$DEP.CEO       <- 0
d2$SEC           <- 0
d2$COO           <- 0
d2$TRUST         <- 0
d2$HUM.RES       <- 0
d2$DEP.HEAD      <- 0
d2$MAN           <- 0
d2$DEV           <- 0
d2$TECH          <- 0
d2$COMM          <- 0
d2$OTHER         <- 0
d2$PROJECT       <- 0
d2$LEGAL         <- 0
d2$FACILITIES    <- 0
d2$ADMIN.SUP     <- 0
d2$MED.MAN       <- 0
d2$HEALTH.HUM    <- 0
d2$TRAIN         <- 0
d2$ACADEMIC.MAN  <- 0
d2$PROFESIONAL   <- 0
d2$OTHER.PROF    <- 0
d2$ACADEMIC.PROF <- 0
d2$MED.PROF      <- 0

# d2[ !(d2$TitleTxt2 %in% All.TITLES.VEC) ] %>% nrow


# d2$CFO        [ !(d2$TitleTxt2 %in% All.TITLES.VEC) ] <- NA
# d2$Treasure   [ !(d2$TitleTxt2 %in% All.TITLES.VEC) ] <- NA
# d2$DEP.CEO    [ !(d2$TitleTxt2 %in% All.TITLES.VEC) ] <- NA
# d2$SEC        [ !(d2$TitleTxt2 %in% All.TITLES.VEC) ] <- NA
# d2$COO        [ !(d2$TitleTxt2 %in% All.TITLES.VEC) ] <- NA
# d2$TRUST      [ !(d2$TitleTxt2 %in% All.TITLES.VEC) ] <- NA
# d2$HUM.RES    [ !(d2$TitleTxt2 %in% All.TITLES.VEC) ] <- NA
# d2$DEP.HEAD   [ !(d2$TitleTxt2 %in% All.TITLES.VEC) ] <- NA
# d2$MAN        [ !(d2$TitleTxt2 %in% All.TITLES.VEC) ] <- NA
# d2$DEV        [ !(d2$TitleTxt2 %in% All.TITLES.VEC) ] <- NA
# d2$OTHER      [ !(d2$TitleTxt2 %in% All.TITLES.VEC) ] <- NA


#' 
#' 
#' ### CEO
#' 
#' First, assign individuals to CEO catagory if they had a title that appeared in the 1000 most common titles that we coded to CEO.
#' 
## ------------------------------------------------------------------------
d2$CEO.Prob <- 0

d2$CEO[ d2$TitleTxt2 %in% All.Titles$CEO.Clear] <- 1

d2$CEO.Prob[ d2$TitleTxt2 %in% All.Titles$CEO.Prob] <- 1

# Test <-  d2 %>% filter ( CEO.Prob == 1 & TrustOrDir == 1 & Officer == 0)
# Test$TitleTxt2 %>% table( ) %>% sort( decreasing=T ) %>% as.data.frame( )
# TRUSTEE not CEO

d2$CEO   [ d2$CEO.Prob == 1 & d2$TrustOrDir == 1 & d2$Officer == 0]   <- 0
d2$TRUST [ d2$CEO.Prob == 1 & d2$TrustOrDir == 1 & d2$Officer == 0]   <- 1

# Test <- d2 %>% filter ( CEO.Prob == 1 & TrustOrDir == 0 & Officer == 1)
# Test$TitleTxt2 %>% table( ) %>% sort( decreasing=T ) %>% as.data.frame( )

# CEO Not Trustee

#' 
#'  TrustOrDir          Officer      KeyEmpl      HighComp       Meaning
#' --------------     ----------    ----------   -----------    -----------
#' 1                   0             0            0              Board Probably not CEO
#' 0                   1             0            0              CEO
#' 1                   1             0            0              CEO
#' 0                   0             1            0              Management
#' 0                   0             0            1              Other
#' 
#' 
## ------------------------------------------------------------------------
d2$CEO   [ d2$CEO.Prob == 1 & d2$TrustOrDir == 0 & d2$Officer == 1]   <- 1
d2$CEO[ d2$CEO.Prob == 1 & d2$TrustOrDir == 1 & d2$Officer == 1] <- 1
d2$TRUST[ d2$CEO.Prob == 1 & d2$TrustOrDir == 1 & d2$Officer == 1] <- 1
d2$TRUST[ d2$TitleTxt2 %in% All.Titles$CEO.Board] <- 1
d2$CEO [ d2$TitleTxt2 %in% All.Titles$CEO.Board] <- 1
d2$MAN [ d2$CEO.Prob == 1 & d2$TrustOrDir == 0 & d2$Officer == 0 & d2$KeyEmpl == 1] <- 1
d2$MAN [ d2$CEO.Prob == 1 & d2$TrustOrDir == 0 & d2$Officer == 0 & d2$HighComp == 1] <- 1


#' 
#' Next we want to assign individuals to the CEO catagory if they had one of the most common phrases associated with that title
#' 
## ------------------------------------------------------------------------

d2$CEO [ grepl( " CHIEF EXECUTIVE OFFICER ", d2$TitleTxt2) ] <- 1
d2$CEO [ grepl( " EXECUTIVE DIRECTOR ", d2$TitleTxt2) ] <- 1
d2$CEO [ grepl( " ASSISTANT EXECUTIVE DIRECTOR ", d2$TitleTxt2) ] <- 0



d2$TitleTxt2[ d2$CEO == 1 ] %>% table() %>% sort( decreasing=T ) %>% as.data.frame( ) %>% head( 20 ) %>% pander


#' 
#' ### HEAD OF SCHOOL
#' 
#'  TrustOrDir          Officer      KeyEmpl      HighComp       Meaning
#' --------------     ----------    ----------   -----------    -----------
#' 1                   0             0            0              Board Probably not CFO?
#' 0                   1             0            0              CFO
#' 1                   1             0            0              CFO
#' 
#' 
## ------------------------------------------------------------------------
d2$TRUST [d2$TitleTxt2 %in% All.Titles$SCHOOL.HEAD & d2$TrustOrDir == 1] <- 1
d2$CEO [d2$TitleTxt2 %in% All.Titles$SCHOOL.HEAD & d2$Officer == 1] <- 1
d2$MAN [d2$TitleTxt2 %in% All.Titles$SCHOOL.HEAD & d2$Officer == 0 & d2$TrustOrDir == 0 & d2$KeyEmpl == 1 ] <- 1
d2$OTHER [d2$TitleTxt2 %in% All.Titles$SCHOOL.HEAD & d2$Officer == 0 & d2$TrustOrDir == 0 & d2$KeyEmpl == 0 ] <- 1


#' 
#' 
#' 
#' 
#' ### CFO
#' 
#' 
#'  TrustOrDir          Officer      KeyEmpl      HighComp       Meaning
#' --------------     ----------    ----------   -----------    -----------
#' 1                   0             0            0              Board Probably not CFO?
#' 0                   1             0            0              CFO
#' 1                   1             0            0              CFO
#' 
#' 
## ------------------------------------------------------------------------

d2$CFO.Prob <- 0

d2$CFO[ d2$TitleTxt2 %in% All.Titles$CFO.Clear] <- 1
d2$CFO.Prob[ d2$TitleTxt2 %in% All.Titles$CFO.Prob] <- 1

d2$CFO   [ d2$CFO.Prob == 1 & d2$TrustOrDir == 1]   <- 0
d2$TRUST [ d2$CFO.Prob == 1 & d2$Officer == 1   ]   <- 1
d2$CFO   [ d2$CFO.Prob == 1 & d2$KeyEmpl == 1]      <- 1
d2$CFO   [ d2$CFO.Prob == 1 & d2$HighComp == 1]     <- 1
d2$CFO   [ grepl( " CFO ", d2$TitleTxt2) ]          <- 1
d2$CFO   [ grepl( " CHIEF FINANCIAL OFFICER ", d2$TitleTxt2) ]          <- 1
d2$CFO   [ grepl( " FINANCE ", d2$TitleTxt2) ]          <- 1

d2$TRUST[ d2$CFO.Prob == 1 & d2$TrustOrDir == 1 & d2$Officer == 1] <- 1

titles <- d2$TitleTxt2[ d2$CFO == 1]


table( titles ) %>% sort( decreasing=T ) %>% as.data.frame( ) %>% head( 20 ) %>% pander


#' 
#' ### Treasurer
#' 
## ------------------------------------------------------------------------

d2$Treasurer  [ d2$TitleTxt2 %in% All.Titles$Treasurer] <- 1
d2$Treasurer  [ grepl( " TREASURER ", d2$TitleTxt2) ]   <- 1


titles <- d2$TitleTxt2[ d2$Treasurer == 1]
table( titles ) %>% sort( decreasing=T ) %>% as.data.frame( ) %>% head( 20 ) %>% pander



#' 
#' ### Deputy CEO
#' 
## ------------------------------------------------------------------------

d2$DEP.CEO[ d2$TitleTxt2 %in% All.Titles$DEP.CEO ] <- 1

d2$DEP.CEO[ d2$TitleTxt2 %in% All.Titles$DEP.CEO.Prob & d2$Officer == 1 ] <- 1

#d2[ d2$TitleTxt2 %in% All.Titles$DEP.CEO.Prob, ]

d2$MAN [ d2$TitleTxt2 %in% All.Titles$DEP.CEO.Prob  & d2$Officer == 0 & (d2$KeyEmpl == 1 | d2$HighComp == 1 ) ] <- 1

titles <- d2$TitleTxt2[ d2$DEP.CEO == 1]
table( titles ) %>% sort( decreasing=T ) %>% as.data.frame( ) %>% head( 20 ) %>% pander



#' 
#' ### Secretary
#' 
## ------------------------------------------------------------------------


d2$SEC[ d2$TitleTxt2 %in%  All.Titles$SEC] <- 1
d2$SEC [ grepl( " SECRETARY ", d2$TitleTxt2) ] <- 1
d2$SEC [ grepl( " SEC ", d2$TitleTxt2) ] <- 1

titles <- d2$TitleTxt2[ d2$SEC == 1]
table( titles ) %>% sort( decreasing=T ) %>% as.data.frame( ) %>% head( 20 ) %>% pander

#' 
#' ### COO
#' 
## ------------------------------------------------------------------------

#All.Titles$COO.Prob

d2$COO[ d2$TitleTxt2 %in%  All.Titles$COO] <- 1
d2$COO[ d2$TitleTxt2 %in%  All.Titles$COO] <- 1
d2$COO [ grepl( " OPERATIONS ", d2$TitleTxt2) ] <- 1
d2$COO [ grepl( " COO ", d2$TitleTxt2) ] <- 1


#d2$COO.Prob[ d2$TitleTxt2 %in% All.Titles$COO.Prob] <- 1
#d2$COO.Prob[ d2$TitleTxt2 %in% All.Titles$COO.Prob] <- 1

#COO.Prob.Names <-d2 %>% filter (COO.Prob == 1 & d2$Officer == 1)

#d2$COO[ COO.Prob == 1] <- 1

titles <- d2$TitleTxt2[ d2$COO == 1]
table( titles ) %>% sort( decreasing=T ) %>% as.data.frame( ) %>% head( 20 ) %>% pander




#' 
#' ### Director/Trustee
#' 
## ------------------------------------------------------------------------

d2$TRUST[ d2$TitleTxt2 %in%  All.Titles$TRUST] <- 1
d2$TRUST [ grepl( " TRUSTEE ", d2$TitleTxt2) ] <- 1
d2$TRUST [ grepl( " TRUST ", d2$TitleTxt2) ] <- 1
d2$TRUST [ grepl( " BOARD MEMBER ", d2$TitleTxt2) ] <- 1
d2$TRUST [ grepl( " CHAIR ", d2$TitleTxt2) ] <- 1
d2$TRUST [ d2$TrustOrDir == 1 ] <- 1

titles <- d2$TitleTxt2[ d2$TRUST == 1]
table( titles ) %>% sort( decreasing=T ) %>% as.data.frame( ) %>% head( 20 ) %>% pander

#' 
#' ### Human Resources
#' 
## ------------------------------------------------------------------------

d2$HUM.RES[ d2$TitleTxt2 %in%  All.Titles$HUM.RES] <- 1

d2$HUM.RES [ grepl( " HUMAN RESOURCES ", d2$TitleTxt2) ] <- 1
d2$HUM.RES [ grepl( " HUMAN RESOURCE ", d2$TitleTxt2) ] <- 1
d2$HUM.RES [ grepl( " STAFFING ", d2$TitleTxt2) ] <- 1


titles <- d2$TitleTxt2[ d2$HUM.RES == 1]
table( titles ) %>% sort( decreasing=T ) %>% as.data.frame( ) %>% head( 20 ) %>% pander

#' 
#' 
#' ### Communications
#' 
## ------------------------------------------------------------------------

d2$COMM[ d2$TitleTxt2 %in%  All.Titles$COMM] <- 1
d2$COMM [ grepl( " COMMUNICATION ", d2$TitleTxt2) ] <- 1
d2$COMM [ grepl( " COMMUNICATIONS ", d2$TitleTxt2) ] <- 1
d2$COMM [ grepl( " MARKETING ", d2$TitleTxt2) ] <- 1
d2$COMM [ grepl( " EXTERNAL AFFAIRS ", d2$TitleTxt2) ] <- 1
d2$COMM [ grepl( " PUBLIC AFFAIRS ", d2$TitleTxt2) ] <- 1
d2$COMM [ grepl( " EXTERNAL AFFAIRS ", d2$TitleTxt2) ] <- 1
d2$COMM [ grepl( " RELATIONS ", d2$TitleTxt2) ] <- 1

titles <- d2$TitleTxt2[ d2$COMM == 1]
table( titles ) %>% sort( decreasing=T ) %>% as.data.frame( ) %>% head( 20 ) %>% pander


#' 
#' 
#' ### Management
#' 
## ------------------------------------------------------------------------

d2$MAN[ d2$TitleTxt2 %in%  All.Titles$MAN] <- 1


titles <- d2$TitleTxt2[ d2$MAN == 1]
table( titles ) %>% sort( decreasing=T ) %>% as.data.frame( ) %>% head( 20 ) %>% pander


#' 
#' ### Technology/Information
#' 
## ------------------------------------------------------------------------
d2$TECH[ d2$TitleTxt2 %in%  All.Titles$TECH] <- 1

d2$TECH [ grepl( " CHIEF INFORMATION OFFICER ", d2$TitleTxt2) ] <- 1
d2$TECH [ grepl( " CIO ", d2$TitleTxt2) ] <- 1
d2$TECH [ grepl( " IT ", d2$TitleTxt2) ] <- 1
d2$TECH [ grepl( " INFORMATION ", d2$TitleTxt2) ] <- 1
d2$TECH [ grepl( " TECHNOLOGY ", d2$TitleTxt2) ] <- 1
d2$TECH [ grepl( " CDO ", d2$TitleTxt2) ] <- 1



titles <- d2$TitleTxt2[ d2$TECH == 1]
table( titles ) %>% sort( decreasing=T ) %>% as.data.frame( ) %>% head( 20 ) %>% pander

#' 
#' ### Development
#' 
## ------------------------------------------------------------------------

d2$DEV [ d2$TitleTxt2 %in%  All.Titles$DEV]       <- 1
d2$DEV [ grepl( " DEVELOPMENT ", d2$TitleTxt2) ]  <- 1
d2$DEV [ grepl( " PHILANTHROPY ", d2$TitleTxt2) ] <- 1
d2$DEV [ grepl( " FUND ", d2$TitleTxt2) ]         <- 1
d2$DEV [ grepl( " FUNDRAISING ", d2$TitleTxt2) ]  <- 1
d2$DEV [ grepl( " MAJOR GIFTS ", d2$TitleTxt2) ]  <- 1


titles <- d2$TitleTxt2[ d2$DEV == 1]
table( titles ) %>% sort( decreasing=T ) %>% as.data.frame( ) %>% head( 20 ) %>% pander



#' 
#' 
#' ### OTHER
#' 
## ------------------------------------------------------------------------


d2$OTHER [ d2$TitleTxt2 %in%  All.Titles$OTHER] <- 1
titles <- d2$TitleTxt2[ d2$OTHER == 1]
table( titles ) %>% sort( decreasing=T ) %>% as.data.frame( ) %>% head( 20 ) %>% pander

#' 
#' 
#' ### PROJECT
#' 
## ------------------------------------------------------------------------

d2$PROJECT [ d2$TitleTxt2 %in%  All.Titles$PROJECT] <- 1

d2$PROJECT [ grepl( " PROJECT ", d2$TitleTxt2) ]  <- 1
d2$PROJECT [ grepl( " PROJECTS ", d2$TitleTxt2) ]  <- 1
d2$PROJECT [ grepl( " PROGRAM ", d2$TitleTxt2) ] <- 1
d2$PROJECT [ grepl( " PROGRAMS ", d2$TitleTxt2) ] <- 1



titles <- d2$TitleTxt2[ d2$PROJECT == 1]
table( titles ) %>% sort( decreasing=T ) %>% as.data.frame( ) %>% head( 20 ) %>% pander

#' 
#' ### LEGAL
#' 
## ------------------------------------------------------------------------

d2$LEGAL [ d2$TitleTxt2 %in%  All.Titles$LEGAL] <- 1
titles <- d2$TitleTxt2[ d2$LEGAL == 1]

d2$LEGAL [ grepl( " COUNSEL ", d2$TitleTxt2) ]  <- 1
d2$LEGAL [ grepl( " ATTORNEY ", d2$TitleTxt2) ]  <- 1
d2$LEGAL [ grepl( " COMPLIANCE ", d2$TitleTxt2) ] <- 1
d2$LEGAL [ grepl( " POLICY ", d2$TitleTxt2) ] <- 1
d2$LEGAL [ grepl( " LEGAL ", d2$TitleTxt2) ] <- 1
d2$LEGAL [ grepl( " LITIGATION ", d2$TitleTxt2) ] <- 1



table( titles ) %>% sort( decreasing=T ) %>% as.data.frame( ) %>% head( 20 ) %>% pander

#' 
#' ### FACILITIES
#' 
## ------------------------------------------------------------------------

d2$FACILITIES [ d2$TitleTxt2 %in%  All.Titles$FACILITIES] <- 1
titles <- d2$TitleTxt2[ d2$FACILITIES == 1]

d2$FACILITIES [ grepl( " FACILITIES ", d2$TitleTxt2) ]  <- 1
d2$FACILITIES [ grepl( " MAINTENANCE ", d2$TitleTxt2) ]  <- 1
d2$FACILITIES [ grepl( " FIELD ", d2$TitleTxt2) ] <- 1
d2$FACILITIES [ grepl( " FIELDS ", d2$TitleTxt2) ] <- 1
d2$FACILITIES [ grepl( " FACILITY ", d2$TitleTxt2) ] <- 1

table( titles ) %>% sort( decreasing=T ) %>% as.data.frame( ) %>% head( 20 ) %>% pander

#' 
#' ### ADMIN.SUP
#' 
## ------------------------------------------------------------------------

d2$ADMIN.SUP [ d2$TitleTxt2 %in%  All.Titles$ADMIN.SUP] <- 1

d2$ADMIN.SUP [ grepl( " SUPPORT ", d2$TitleTxt2)  & grepl( " SERVICES ", d2$TitleTxt2) ] <- 1
d2$ADMIN.SUP [ grepl( " ADMINISTRATION ", d2$TitleTxt2)  & grepl( " ASSISTANT ", d2$TitleTxt2) ] <- 1
d2$ADMIN.SUP [ grepl( " ADMINISTRATIVE ", d2$TitleTxt2)  & grepl( " ASSISTANT ", d2$TitleTxt2) ] <- 1
d2$ADMIN.SUP [ grepl( " EXECUTIVE ", d2$TitleTxt2)  & grepl( " ASSISTANT ", d2$TitleTxt2) ] <- 1

titles <- d2$TitleTxt2[ d2$ADMIN.SUP == 1]
table( titles ) %>% sort( decreasing=T ) %>% as.data.frame( ) %>% head( 20 ) %>% pander

#' 
#' ### MED.MAN
#' 
## ------------------------------------------------------------------------

d2$MED.MAN [ d2$TitleTxt2 %in%  All.Titles$MED.MAN] <- 1
titles <- d2$TitleTxt2[ d2$MED.MAN == 1]
table( titles ) %>% sort( decreasing=T ) %>% as.data.frame( ) %>% head( 20 ) %>% pander

#' 
#' ### HEALTH.HUM
#' 
## ------------------------------------------------------------------------

d2$HEALTH.HUM [ d2$TitleTxt2 %in%  All.Titles$HEALTH.HUM] <- 1

d2$HEALTH.HUM[ grepl( " CASE ", d2$TitleTxt2)] <- 1

d2$HEALTH.HUM[ grepl( " SHELTER ", d2$TitleTxt2)] <- 1
d2$HEALTH.HUM[ grepl( " HOUSING ", d2$TitleTxt2)] <- 1
d2$HEALTH.HUM[ grepl( " SOCIAL WORKER ", d2$TitleTxt2)] <- 1


titles <- d2$TitleTxt2[ d2$HEALTH.HUM == 1]
table( titles ) %>% sort( decreasing=T ) %>% as.data.frame( ) %>% head( 20 ) %>% pander

#' 
#' ### TRAIN
#' 
## ------------------------------------------------------------------------

d2$TRAIN [ d2$TitleTxt2 %in%  All.Titles$TRAIN] <- 1
titles <- d2$TitleTxt2[ d2$TRAIN == 1]
table( titles ) %>% sort( decreasing=T ) %>% as.data.frame( ) %>% head( 20 ) %>% pander

#' 
#' ### Academic Management
#' 
## ------------------------------------------------------------------------

d2$ACADEMIC.MAN [ d2$TitleTxt2 %in%  All.Titles$ACADEMIC.MAN] <- 1
titles <- d2$TitleTxt2[ d2$ACADEMIC.MAN == 1]
table( titles ) %>% sort( decreasing=T ) %>% as.data.frame( ) %>% head( 20 ) %>% pander

#' 
#' ### DEP.HEAD
#' 
## ------------------------------------------------------------------------

d2$DEP.HEAD [ d2$TitleTxt2 %in%  All.Titles$DEP.HEAD] <- 1
titles <- d2$TitleTxt2[ d2$DEP.HEAD == 1]
table( titles ) %>% sort( decreasing=T ) %>% as.data.frame( ) %>% head( 20 ) %>% pander

#' 
#' ### PROFESIONAL
#' 
## ------------------------------------------------------------------------

d2$PROFESIONAL [ d2$TitleTxt2 %in%  All.Titles$PROFESIONAL] <- 1
titles <- d2$TitleTxt2[ d2$PROFESIONAL == 1]
table( titles ) %>% sort( decreasing=T ) %>% as.data.frame( ) %>% head( 20 ) %>% pander

#' 
#' ### OTHER PROF
#' 
## ------------------------------------------------------------------------

d2$OTHER.PROF [ d2$TitleTxt2 %in%  All.Titles$OTHER.PROF] <- 1
titles <- d2$TitleTxt2[ d2$OTHER.PROF == 1]
table( titles ) %>% sort( decreasing=T ) %>% as.data.frame( ) %>% head( 20 ) %>% pander

#' 
#' ### ACADEMIC PROF
#' 
## ------------------------------------------------------------------------

d2$ACADEMIC.PROF [ d2$TitleTxt2 %in%  All.Titles$ACADEMIC.PROF] <- 1
titles <- d2$TitleTxt2[ d2$ACADEMIC.PROF == 1]
table( titles ) %>% sort( decreasing=T ) %>% as.data.frame( ) %>% head( 20 ) %>% pander

#' 
#' ### MED PROF
#' 
## ------------------------------------------------------------------------
d2$MED.PROF [ d2$TitleTxt2 %in%  All.Titles$MED.PROF] <- 1
titles <- d2$TitleTxt2[ d2$MED.PROF == 1]
table( titles ) %>% sort( decreasing=T ) %>% as.data.frame( ) %>% head( 20 ) %>% pander

#' 
#' 
#' 
#' 
#' 
#' ### Common Missed Catagories.
#' 
## ----HOW MANY------------------------------------------------------------
# d2$Num.Titles <- d2$CEO + d2$CFO + d2$Treasurer + d2$DEP.CEO + d2$SEC + d2$COO + d2$TRUST + d2$HUM.RES + d2$DEP.HEAD + d2$MAN + d2$DEV + d2$OTHER +d2$TECH + d2$COMM
# 
# d2$Catagorized <-(d2$Num.Titles >= 1)
# 
# sum(d2$Catagorized)/nrow(d2)
# 
# d2.missed <- d2[ d2$Catagorized != TRUE,]
# 
# Broken.Tables <- d2.missed$TitleTxt2 %>% table( ) %>% sort( decreasing=T ) %>% as.data.frame( ) %>% head(100)
# Broken.Tables %>% pander()


#" CHIEF PARTY "

d2$CEO [ grepl( " CHIEF PARTY ", d2$TitleTxt2) & d2$Officer == 1 ] <- 1
d2$TRUST [ grepl( " CHIEF PARTY ", d2$TitleTxt2) & d2$FmrOfficer == 1  ] <- 1
d2$MAN [ grepl( " CHIEF PARTY ", d2$TitleTxt2) & d2$Officer == 0  & d2$FmrOfficer == 0 & d2$KeyEmpl == 1] <- 1
d2$OTHER [ grepl( " CHIEF PARTY ", d2$TitleTxt2) & d2$Officer == 0  & d2$FmrOfficer == 0 & d2$KeyEmpl ==  0] <- 1


d2$OTHER [ grepl( " FORMER KEY EMPLOYEE ", d2$TitleTxt2) ] <- 1
d2$OTHER [ grepl( " ASSOCIATE PROFESSOR ", d2$TitleTxt2) ] <- 1
d2$MAN [ grepl( " FORMER DIRECTOR ", d2$TitleTxt2) ] <- 1
d2$OTHER [ grepl( " ASSISTANT PROFESSOR ", d2$TitleTxt2) ] <- 1
d2$MAN [ grepl( " FORMER VICE PRESIDENT ", d2$TitleTxt2) ] <- 1
d2$DEP.HEAD [ grepl( " DIRECTOR OF NURSING ", d2$TitleTxt2) ] <- 1
d2$DEP.HEAD [ grepl( " VICE PRESIDENT PROFESSIONAL SERVICES ", d2$TitleTxt2) ] <- 1
d2$DEP.HEAD [ grepl( " CHIEF PROFESSIONAL OFFICER ", d2$TitleTxt2) ] <- 1

d2$OTHER [ grepl( " LAW PROFESSOR ", d2$TitleTxt2) & d2$TrustOrDir == 0 ] <- 1
d2$TRUST [ grepl( " LAW PROFESSOR ", d2$TitleTxt2) & d2$TrustOrDir == 1] <- 1

d2$CEO [ grepl( " FORMER HEAD SCHOOL ", d2$TitleTxt2) & d2$Officer == 1 ] <- 1
d2$TRUST [ grepl( " FORMER HEAD SCHOOL ", d2$TitleTxt2) & d2$FmrOfficer == 1  ] <- 1
d2$MAN [ grepl( " FORMER HEAD SCHOOL ", d2$TitleTxt2) & d2$Officer == 0  & d2$FmrOfficer == 0 ] <- 1



d2$CEO [ grepl( " CEO FORMER ", d2$TitleTxt2) ] <- 1
d2$MAN [ grepl( " DEAN PROFESSOR ", d2$TitleTxt2) ] <- 1
d2$MAN [ grepl( " FORMER EXECUTIVE VICE PRESIDENT ", d2$TitleTxt2) ] <- 1
d2$DEP.CEO [ grepl( " ASSISTANT CEO ", d2$TitleTxt2) ] <- 1

d2$OTHER [ grepl( " PHARMACY ", d2$TitleTxt2) ] <- 1
d2$OTHER [ grepl( " PHARMACY ", d2$TitleTxt2) ] <- 1

d2$MAN [ grepl( " REGIONAL MANAGER ", d2$TitleTxt2) ] <- 1
d2$DEP.HEAD [ grepl( " RN CASE MANAGER ", d2$TitleTxt2) ] <- 1

d2$OTHER [ grepl( " SOCIAL WORKER ", d2$TitleTxt2) & d2$TrustOrDir == 0 ] <- 1
d2$TRUST [ grepl( " SOCIAL WORKER ", d2$TitleTxt2) & d2$TrustOrDir == 1] <- 1

d2$DEP.HEAD [ grepl( " VICE PRESIDENT PATIENT CARE CNO ", d2$TitleTxt2) ] <- 1

d2$OTHER [ grepl( " CONCERTMASTER ", d2$TitleTxt2) ] <- 1
d2$OTHER [ grepl( " CONCERTMASTER ", d2$TitleTxt2) ] <- 1

d2$CEO [ grepl( " HS PRINCIPAL ", d2$TitleTxt2) & d2$Officer == 1 ] <- 1
d2$TRUST [ grepl( " HS PRINCIPAL ", d2$TitleTxt2) & d2$FmrOfficer == 1  ] <- 1
d2$MAN [ grepl( " HS PRINCIPAL ", d2$TitleTxt2) & d2$Officer == 0  & d2$FmrOfficer == 0 ] <- 1

d2$MAN [ grepl( " SENIOR MANAGER ", d2$TitleTxt2) ] <- 1
d2$DEP.HEAD [ grepl( " VICE PRESIDENT CAO ", d2$TitleTxt2) ] <- 1
d2$LEGAL [ grepl( " VICE PRESIDENT LEGAL ", d2$TitleTxt2) ] <- 1



#' 
#' 
#' 
#' ## Total Number of Titles Catagorized
#' 
## ------------------------------------------------------------------------
d2$Num.Titles <- d2$CEO + d2$CFO + d2$Treasurer + d2$DEP.CEO + d2$SEC + d2$COO + d2$TRUST + d2$HUM.RES + d2$DEP.HEAD + d2$MAN + d2$DEV + d2$OTHER +d2$TECH + d2$COMM + d2$PROJECT + d2$LEGAL + d2$FACILITIES + d2$ADMIN.SUP + d2$MED.MAN + d2$HEALTH.HUM + d2$TRAIN + d2$ACADEMIC.MAN + d2$PROFESIONAL + d2$OTHER.PROF + d2$ACADEMIC.PROF + d2$MED.PROF

d2$Catagorized <-(d2$Num.Titles >= 1)

d2.missed <- d2[ d2$Catagorized != TRUE,]
cat("\nAfter Standardizing, cleaning, and categorizing all of the 50 most Common titles, we Categorized ", sum(d2$Catagorized), " of ", nrow(d2) , " titles. \n"  )
cat("This accounted for ", sum(d2$Catagorized)/nrow(d2), " of the Observations\n")

#' 
#' 
#' 
#' ## Catagory Descriptive Statistics
#' 
## ------------------------------------------------------------------------
n<-nrow(d2)
Cat.Desc <- matrix(2,2,2) %>% as.data.frame()

Cat.Desc[1,] <- c(sum(d2$CEO), sum(d2$CEO)/n)
Cat.Desc[2,] <- c(sum(d2$CFO), sum(d2$CFO)/n)
Cat.Desc[3,] <- c(sum(d2$Treasurer), sum(d2$Treasurer)/n)
Cat.Desc[4,] <- c(sum(d2$DEP.CEO), sum(d2$DEP.CEO)/n)
Cat.Desc[5,] <- c(sum(d2$SEC), sum(d2$SEC)/n)
Cat.Desc[6,] <- c(sum(d2$COO), sum(d2$COO)/n)
Cat.Desc[7,] <- c(sum(d2$TRUST), sum(d2$TRUST)/n)
Cat.Desc[8,] <- c(sum(d2$HUM.RES), sum(d2$HUM.RES)/n)
Cat.Desc[9,] <- c(sum(d2$DEP.HEAD), sum(d2$DEP.HEAD)/n)
Cat.Desc[10,] <- c(sum(d2$MAN), sum(d2$MAN)/n)
Cat.Desc[11,] <- c(sum(d2$DEV), sum(d2$DEV)/n)
Cat.Desc[12,] <- c(sum(d2$TECH), sum(d2$TECH)/n)
Cat.Desc[13,] <- c(sum(d2$COMM), sum(d2$COMM)/n)
Cat.Desc[14,] <- c(sum(d2$PROJECT), sum(d2$PROJECT)/n)
Cat.Desc[15,] <- c(sum(d2$LEGAL), sum(d2$LEGAL)/n)
Cat.Desc[16,] <- c(sum(d2$FACILITIES), sum(d2$FACILITIES)/n)
Cat.Desc[17,] <- c(sum(d2$ADMIN.SUP), sum(d2$ADMIN.SUP)/n)
Cat.Desc[18,] <- c(sum(d2$MED.MAN), sum(d2$MED.MAN)/n)
Cat.Desc[19,] <- c(sum(d2$HEALTH.HUM), sum(d2$HEALTH.HUM)/n)
Cat.Desc[20,] <- c(sum(d2$TRAIN), sum(d2$TRAIN)/n)
Cat.Desc[21,] <- c(sum(d2$ACADEMIC.MAN), sum(d2$ACADEMIC.MAN)/n)
Cat.Desc[22,] <- c(sum(d2$PROFESIONAL), sum(d2$PROFESIONAL)/n)
Cat.Desc[23,] <- c(sum(d2$OTHER.PROF), sum(d2$OTHER.PROF)/n)
Cat.Desc[24,] <- c(sum(d2$ACADEMIC.PROF), sum(d2$ACADEMIC.PROF)/n)
Cat.Desc[25,] <- c(sum(d2$MED.PROF), sum(d2$MED.PROF)/n)
Cat.Desc[26,] <- c(sum(d2$OTHER), sum(d2$OTHER)/n)
Cat.Desc[27,] <- c(sum(d2$Catagorized), sum(d2$Catagorized)/n)

Cats <-c("CEO","CFO","Treasuer", "Deputy CEO ", "Secretary", "COO", "Trustee or BoD", "Human Resources", "Department Head", "Generic Management", "Development", "Technology", "Communications", "Project", "Legal", "Facilities", "Administrative Support", "Medical Management", "Health and Human Services", "Training", "Academic Management", "Professional", "Other Profession", "Academic Professional", "Medical Professional", "Other", "Total")
Cat.Desc<-cbind(Cats, Cat.Desc)
colnames(Cat.Desc) <- c("Catagories", "Total", "Percentage of obs")
Cat.Desc %>% pander()



#' 
#' 
#' 
#' 
#' ## Positions and Roles {.tabset}
#' 
#' ![Example of Positions and Roles Assignment](Title Position Roles.png)
#' 
#' 4 Positions:
#' 
#' - Management
#' - HPP
#' - Other Staff
#' - Trustee
#' 
#' 4 Roles
#' 
#' - Board Leadership
#' - C-Level/Officer
#' - Interim 
#' - Former 
#' 
#' 
#' 
#' 
#' ### Management
#' 
#' 
## ------------------------------------------------------------------------
d2$Mgmt <- 0

d2$Mgmt[ d2$CEO == 1 ] <- 1
d2$Mgmt[ d2$CFO == 1 ] <- 1
d2$Mgmt[ d2$DEP.CEO == 1 ] <- 1
d2$Mgmt[ d2$COO == 1 ] <- 1
d2$Mgmt[ d2$HUM.RES == 1 ] <- 1
d2$Mgmt[ d2$DEP.HEAD == 1 ] <- 1
d2$Mgmt[ d2$MAN == 1 ] <- 1
d2$Mgmt[ d2$DEV == 1 ] <- 1
d2$Mgmt[ d2$TECH == 1 ] <- 1
d2$Mgmt[ d2$COMM == 1 ] <- 1
d2$Mgmt[ d2$MED.MAN == 1 ] <- 1
d2$Mgmt[ d2$ACADEMIC.MAN == 1 ] <- 1


#' 
#' 
#' 
#' 
#' ### HPP
#' 
## ------------------------------------------------------------------------
d2$HPP <- 0

d2$HPP[ d2$PROFESIONAL == 1 ] <- 1
d2$HPP[ d2$LEGAL == 1 ] <- 1
d2$HPP[ d2$PROJECT == 1 ] <- 1
d2$HPP[ d2$TRAIN == 1 ] <- 1


#' 
#' 
#' 
#' ### Other Staff
#' 
## ------------------------------------------------------------------------
d2$Other.Staff <- 0

d2$Other.Staff[ d2$HPP == 0 & d2$Mgmt == 0 & d2$TRUST == 0] <- 1
 


#' 
#' 
#' 
#' ### Trustee
#' 
## ------------------------------------------------------------------------
d2$Trustee <- 0


#' 
#' 
#' 
#' ### Board Leadership
#' 
## ------------------------------------------------------------------------
d2$Board.Leadership <- 0

d2$Board.Leadership[ d2$CEO == 1 & d2$TRUST == 1 ] <- 1
d2$Board.Leadership[ d2$COO == 1 & d2$TRUST == 1 ] <- 1
d2$Board.Leadership[ d2$CFO == 1 & d2$TRUST == 1 ] <- 1
d2$Board.Leadership[ d2$Treasurer == 1 & d2$TRUST == 1 ] <- 1
d2$Board.Leadership[ d2$DEP.CEO == 1 & d2$TRUST == 1 ] <- 1
d2$Board.Leadership[ d2$SEC == 1 & d2$TRUST == 1 ] <- 1
d2$Board.Leadership[ d2$Officer == 1 & d2$TRUST == 1 ] <- 1


#' 
#' 
#' 
#' 
#' ### C-Level / Officer
#' 
## ------------------------------------------------------------------------

d2$C.Level <- 0

d2$C.Level [ grepl( " C[A-Z]O ", d2$TitleTxt2) ] <- 1
d2$C.Level [ grepl( " CHIEF ", d2$TitleTxt2) ] <- 1
d2$C.Level [ d2$Officer == 1 ] <- 1

#' 
#' 
#' ### Interim
#' 
## ------------------------------------------------------------------------
d2$Interim <- 0

d2$Interim [ grepl( "SINCE", toupper(d2$TitleTxt) ) ] <- 1
d2$Interim [ grepl( "FROM", toupper(d2$TitleTxt) ) ] <- 1
d2$Interim [ grepl( "INTERIM", toupper(d2$TitleTxt) ) ] <- 1
d2$Interim [ grepl( "TEMP", toupper(d2$TitleTxt) ) ] <- 1
d2$Interim [ grepl( "TEMPORARY", toupper(d2$TitleTxt) ) ] <- 1


#' 
#' 
#' ### Former
#' 
## ------------------------------------------------------------------------
d2$Former <- 0
d2$Former [ grepl( "FORMER", toupper(d2$TitleTxt) ) ] <- 1
d2$Former [ grepl( "THRU", toupper(d2$TitleTxt) ) ] <- 1
d2$Former [ grepl( "THROUGH", toupper(d2$TitleTxt) ) ] <- 1
d2$Former [ grepl( "PAST", toupper(d2$TitleTxt) ) ] <- 1
d2$Former [ d2$FmrOfficer == 1 ] <- 1


#' 
#' 
#' 
#' ## Roles and Positions Descriptive Statistics
#' 
## ------------------------------------------------------------------------
n <- nrow(d2)
RnP.Desc <- matrix(2,2,2) %>% as.data.frame()

RnP.Desc[1,] <- c(sum(d2$Mgmt), sum(d2$Mgmt)/n)
RnP.Desc[2,] <- c(sum(d2$HPP), sum(d2$HPP)/n)
RnP.Desc[3,] <- c(sum(d2$Other.Staff), sum(d2$Other.Staff)/n)
RnP.Desc[4,] <- c(sum(d2$Board.Leadership), sum(d2$Board.Leadership)/n)
RnP.Desc[5,] <- c(sum(d2$C.Level), sum(d2$C.Level)/n)
RnP.Desc[6,] <- c(sum(d2$Interim), sum(d2$Interim)/n)
RnP.Desc[7,] <- c(sum(d2$Former), sum(d2$Former)/n)

Cats <- c("Mgmt", "HPP", "Other Staff","Board Leadership", "C-Level", "Interim", "Former")
RnP.Desc <- cbind(Cats, RnP.Desc)
colnames(RnP.Desc) <- c("Roles and Positions", "Total", "Percentage of obs")
RnP.Desc %>% pander()


#' 
#' 
## ------------------------------------------------------------------------
saveRDS( d2, "cleaned-titles/categorized-titles.rds")
